refactor(view): move the eleven keymap sections into lua/dadbod-grip/view/
view.lua was 5377 lines, of which 2887 were the ten keymap sections plus the
column-highlight autocmd that b4797a3 had split out of _setup_keymaps. Split
functions in one file only fix the nesting, not the size: the file was still
the place every grid feature had to be edited, and reviewing a change to the
FK mappings meant scrolling past the JSON tree viewer.

Each section is now a module exporting setup(bufnr, ctx), the signature the
section functions already had. Ten of the eleven bodies moved byte for byte;
only keymaps_tool changed, and only where it iterated the session registry.

The section modules must not require view.lua, because view.lua requires them:
that would be a cycle. So everything they used from the file arrives on ctx.
ctx.view is the module table itself, which is what the 72 M.* call sites now
resolve through, including M._column_set, whose export and signature a test
pins. The file-local render helpers (open_info_float, resolve_row_bp,
reveal_col_edge, the badge and watch-timer helpers, the augroup and the
column-highlight namespace) are injected the same way; none of them is
reassigned after load, so passing the value rather than an accessor is safe.
ctx.session() and ctx.is_editable() stay functions for the reason d8123d2
gave: a re-query replaces the session wholesale.

gL and gJ iterated pairs(M._sessions), unreachable from another file. They go
through ctx.each_session(fn) instead of a reference to the registry, so no
section names the private table. In gJ the loop became a callback and its
`goto continue` became a `return`.

_setup_keymaps now walks KEYMAP_SECTIONS. Registration order is the precedence
rule for a duplicated lhs, so the list keeps the exact order of the old call
sequence; nvim_buf_get_keymap over every mode on a grid opened against
tests/seed_sqlite.db yields the same 115 mappings, desc for desc, with an
empty diff against a worktree at d8123d2.
