+++ lua/dadbod-grip/view.lua
@@ -2118,10 +2118,30 @@ local function resolve_row_bp(r, line, fallback)
 end
 
 -- ── keymap wiring ─────────────────────────────────────────────────────────
--- One function per keymap group; they all share the helper set built once by
--- make_keymap_ctx() instead of redefining it. Registration order matters: for
--- a given lhs the last vim.keymap.set() wins, so _setup_keymaps() must keep
--- calling the sections in the order they are defined below.
+-- One module per keymap group under view/, each exporting setup(bufnr, ctx);
+-- they all share the helper set built once by make_keymap_ctx() instead of
+-- redefining it. Registration order matters: for a given lhs the last
+-- vim.keymap.set() wins, so _setup_keymaps() walks KEYMAP_SECTIONS in order.
+--
+-- The section modules never require view.lua: this file requires them, so the
+-- reverse would be a cycle. Anything they need from here rides on ctx.
+
+--- The keymap sections, in registration order. This list IS the precedence
+--- rule for a duplicated lhs, not a stylistic ordering: reordering it silently
+--- changes which mapping survives. Append, do not shuffle.
+local KEYMAP_SECTIONS = {
+  require("dadbod-grip.view.keymaps_edit"),
+  require("dadbod-grip.view.keymaps_visual_batch"),
+  require("dadbod-grip.view.keymaps_inspect"),
+  require("dadbod-grip.view.keymaps_nav"),
+  require("dadbod-grip.view.keymaps_misc"),
+  require("dadbod-grip.view.keymaps_sort_filter"),
+  require("dadbod-grip.view.keymaps_fk"),
+  require("dadbod-grip.view.keymaps_aggregate"),
+  require("dadbod-grip.view.keymaps_tool"),
+  require("dadbod-grip.view.keymaps_tab_view"),
+  require("dadbod-grip.view.column_highlight"),
+}
 
 --- Helpers shared by every keymap section: the four map wrappers, the visual
 --- selection row collector and the cell editor (also reached via <CR>).
@@ -2140,6 +2160,30 @@ local function make_keymap_ctx(bufnr)
   function ctx.is_editable()
     return is_editable(M._sessions[bufnr])
   end
+
+  -- The view module itself. Handed over rather than require()d by the section
+  -- modules: view.lua requires them at load time, so a require back would be a
+  -- cycle. Injecting the table keeps the sections free of any view import.
+  ctx.view = M
+
+  -- gL/gJ need every live grid, not just this buffer's. An iterator instead of
+  -- the registry table so no section has to name M._sessions.
+  function ctx.each_session(fn)
+    for b, s in pairs(M._sessions) do fn(b, s) end
+  end
+
+  -- File-local render/buffer helpers the sections call. Injected for the same
+  -- reason as ctx.view: they are not part of the module's public surface.
+  ctx.open_info_float = open_info_float
+  ctx.resolve_row_bp  = resolve_row_bp
+  ctx.reveal_col_edge = reveal_col_edge
+  ctx.update_badge    = _update_badge
+  ctx.start_watch     = _start_watch
+  ctx.stop_watch      = _stop_watch
+  ctx.augroup         = _ag
+  ctx.col_hl_ns       = _col_hl_ns
+  ctx.NULL_DISPLAY    = NULL_DISPLAY
+  ctx.VIEW_LABELS     = VIEW_LABELS
 
   function ctx.map(key, fn, desc)
     vim.keymap.set("n", key, fn, { buffer = bufnr, desc = desc, silent = true })
@@ -2194,2890 +2238,13 @@ local function make_keymap_ctx(bufnr)
 
   return ctx
 end
-
-
---- Row and cell editing: welcome/query pad entry, refresh, edit, delete,
---- insert, clone, apply, undo/redo, yank, NULL, column set.
-local function setup_edit_keymaps(bufnr, ctx)
-  local kmap, edit_cell = ctx.kmap, ctx.edit_cell
-  -- Q: go to welcome screen (home)
-  kmap("welcome", function() require("dadbod-grip").open_welcome() end, "Welcome screen")
-
-  -- q: open query pad (pre-filled with current query)
-  kmap("query_pad", function()
-    local query_pad = require("dadbod-grip.query_pad")
-    local session_q = ctx.session()
-    local s_url = session_q and session_q.url
-    local initial_sql
-    if session_q and session_q.query_spec then
-      initial_sql = qmod.clean_sql(session_q.query_spec)
-    elseif session_q and session_q.query_sql then
-      initial_sql = session_q.query_sql
-    end
-    query_pad.open(s_url, initial_sql and { initial_sql = initial_sql } or nil)
-  end, "Open query pad")
-
-  -- r: refresh
-  kmap("grid_refresh", function()
-    local session = ctx.session()
-    if not session then return end
-    if data.has_changes(session.state) then
-      local staged = data.count_staged(session.state)
-      local choice = vim.fn.confirm(
-        string.format("Refresh will discard %d unapplied change(s). Continue?", staged),
-        "&Yes\n&Cancel", 2
-      )
-      if choice ~= 1 then return end
-    end
-    if session.on_refresh then session.on_refresh(bufnr) end
-  end, "Refresh query")
-
-  kmap("grid_edit", edit_cell, "Edit cell")
-
-  -- d: toggle delete row
-  kmap("grid_delete", function()
-    local session = ctx.session()
-    if not session then return end
-    if not ctx.is_editable() then
-      vim.notify("Read-only: no primary key detected", vim.log.levels.INFO)
-      return
-    end
-    local cell = M.get_cell(bufnr)
-    if not cell then
-      vim.notify("Move cursor to a data row", vim.log.levels.INFO)
-      return
-    end
-    if session.on_delete then session.on_delete(bufnr, cell.row_idx) end
-  end, "Toggle delete row")
-
-  -- o: insert new row
-  kmap("grid_insert", function()
-    local session = ctx.session()
-    if not session then return end
-    if not ctx.is_editable() then
-      vim.notify("Read-only: no primary key detected", vim.log.levels.INFO)
-      return
-    end
-    local cell = M.get_cell(bufnr)
-    local after = cell and cell.row_idx or #session.state.rows
-    if session.on_insert then session.on_insert(bufnr, after) end
-  end, "Insert row after cursor")
-
-  -- c: clone current row (staged INSERT with copied values, PKs cleared)
-  kmap("grid_clone", function()
-    local session = ctx.session()
-    if not session then return end
-    if not ctx.is_editable() then
-      vim.notify("Read-only: no primary key detected", vim.log.levels.INFO)
-      return
-    end
-    local cell = M.get_cell(bufnr)
-    if not cell then
-      vim.notify("Move cursor to a data row to clone", vim.log.levels.INFO)
-      return
-    end
-    if session.on_clone then session.on_clone(bufnr, cell.row_idx) end
-  end, "Clone row (copy values, clear PKs)")
-
-  -- a: apply all staged changes
-  kmap("grid_apply", function()
-    local session = ctx.session()
-    if not session then return end
-
-    -- Mutation preview mode: execute the pending mutation
-    if session.pending_mutation then
-      local pm = session.pending_mutation
-      local choice = vim.fn.confirm(
-        string.format("Execute %s? (%d row%s affected)\n\n%s",
-          pm.type, pm.row_count, pm.row_count == 1 and "" or "s",
-          pm.sql:sub(1, 200)),
-        "&Execute\n&Cancel", 2)
-      if choice ~= 1 then return end
-      local t0 = vim.uv.hrtime()
-      local _, err = db.execute(pm.sql, session.url)
-      local ms = math.floor((vim.uv.hrtime() - t0) / 1e6)
-      if err then
-        vim.notify("Failed: " .. err, vim.log.levels.ERROR)
-        return
-      end
-      vim.notify(string.format("%s executed (%dms, %d row%s)",
-        pm.type, ms, pm.row_count, pm.row_count == 1 and "" or "s"), vim.log.levels.INFO)
-      local history = require("dadbod-grip.history")
-      history.record({ sql = pm.sql, url = session.url, type = pm.type:lower(), elapsed_ms = ms })
-      -- Clear mutation state and reopen the full table
-      local tbl = pm.table_name
-      local s_url = session.url
-      session.pending_mutation = nil
-      session._mutation_title = nil
-      -- Reopen the full table (not the WHERE-filtered preview)
-      local current_win = vim.api.nvim_get_current_win()
-      local grip = require("dadbod-grip")
-      grip.open(tbl, s_url, { reuse_win = current_win })
-      return
-    end
-
-    -- Normal mode: apply staged changes
-    if not ctx.is_editable() then
-      vim.notify("Read-only: no primary key detected", vim.log.levels.INFO)
-      return
-    end
-    if not data.has_changes(session.state) then
-      vim.notify("No staged changes", vim.log.levels.INFO)
-      return
-    end
-    local staged = data.count_staged(session.state)
-    local choice = vim.fn.confirm(
-      string.format("Apply %d staged change(s) to database?", staged),
-      "&Apply\n&Cancel", 2
-    )
-    if choice ~= 1 then return end
-    if session.on_apply then session.on_apply(bufnr) end
-  end, "Apply staged changes")
-
-  -- u: two-tier undo
-  -- Tier 0: mutation preview cancel
-  -- Tier 1: local staging undo (uncommitted changes)
-  -- Tier 2: transaction undo (reverse committed SQL)
-  kmap("grid_undo", function()
-    local session = ctx.session()
-    if not session then return end
-
-    -- Tier 0: mutation preview: cancel (close the preview)
-    if session.pending_mutation then
-      session.pending_mutation = nil
-      session._mutation_title = nil
-      vim.notify("Mutation cancelled", vim.log.levels.INFO)
-      -- Close the preview grid window
-      local win = vim.fn.bufwinid(bufnr)
-      if win ~= -1 then
-        pcall(vim.api.nvim_win_close, win, true)
-      end
-      return
-    end
-
-    -- Tier 1: local staging undo
-    if session._undo_stack and #session._undo_stack > 0 then
-      if not session._redo_stack then session._redo_stack = {} end
-      table.insert(session._redo_stack, session.state)
-      local prev_state = table.remove(session._undo_stack)
-      M.render(bufnr, prev_state)
-      local remaining = #session._undo_stack
-      if remaining > 0 then
-        vim.notify("Undo (" .. remaining .. " more)", vim.log.levels.INFO)
-      else
-        vim.notify("Undo (back to original)", vim.log.levels.INFO)
-      end
-      return
-    end
-
-    -- Tier 2: transaction undo
-    if session._txn_undo_stack and #session._txn_undo_stack > 0 then
-      local reverse = session._txn_undo_stack[#session._txn_undo_stack]
-      local count = #reverse
-      local choice = vim.fn.confirm(
-        "Undo last committed transaction? (" .. count .. " statement(s))",
-        "&Yes\n&No", 2)
-      if choice ~= 1 then return end
-      local txn = "BEGIN;\n" .. table.concat(reverse, ";\n") .. ";\nCOMMIT;"
-      local _, err = db.execute(txn, session.url)
-      if err then
-        vim.notify("Undo failed: " .. err, vim.log.levels.ERROR)
-        return
-      end
-      table.remove(session._txn_undo_stack)
-      vim.notify("Undid transaction (" .. count .. " statement(s))", vim.log.levels.INFO)
-      if session.on_refresh then session.on_refresh(bufnr) end
-      return
-    end
-
-    vim.notify("Nothing to undo", vim.log.levels.INFO)
-  end, "Undo last edit")
-
-  -- U: undo all (resets to original state) or cancel mutation
-  kmap("grid_undo_all", function()
-    local session = ctx.session()
-    if not session then return end
-    -- Mutation preview: cancel
-    if session.pending_mutation then
-      session.pending_mutation = nil
-      session._mutation_title = nil
-      vim.notify("Mutation cancelled", vim.log.levels.INFO)
-      local win = vim.fn.bufwinid(bufnr)
-      if win ~= -1 then pcall(vim.api.nvim_win_close, win, true) end
-      return
-    end
-    if not session._undo_stack or #session._undo_stack == 0 then
-      vim.notify("No staged changes", vim.log.levels.INFO)
-      return
-    end
-    local original = session._undo_stack[1]
-    local count = #session._undo_stack
-    session._undo_stack = {}
-    M.render(bufnr, original)
-    vim.notify("Undid all " .. count .. " change(s)", vim.log.levels.INFO)
-  end, "Undo all changes")
-
-  -- <C-r>: redo
-  kmap("grid_redo", function()
-    local session = ctx.session()
-    if not session then return end
-    if not session._redo_stack or #session._redo_stack == 0 then
-      vim.notify("Nothing to redo", vim.log.levels.INFO)
-      return
-    end
-    if not session._undo_stack then session._undo_stack = {} end
-    table.insert(session._undo_stack, session.state)
-    local redo_state = table.remove(session._redo_stack)
-    M.render(bufnr, redo_state)
-    local remaining = #session._redo_stack
-    vim.notify("Redo" .. (remaining > 0 and " (" .. remaining .. " more)" or ""), vim.log.levels.INFO)
-  end, "Redo")
-
-  -- y: yank cell value
-  kmap("grid_yank_cell", function()
-    local cell = M.get_cell(bufnr)
-    if not cell then
-      vim.notify("Move cursor to a data row to yank", vim.log.levels.INFO)
-      return
-    end
-    local val = cell.value or ""
-    vim.fn.setreg("+", val)
-    vim.notify("Yanked: " .. val, vim.log.levels.INFO)
-  end, "Yank cell value")
-
-  -- Y: yank row as CSV
-  kmap("grid_yank_row", function()
-    local session = ctx.session()
-    if not session then return end
-    local cell = M.get_cell(bufnr)
-    if not cell then
-      vim.notify("Move cursor to a data row to yank", vim.log.levels.INFO)
-      return
-    end
-    local st = session.state
-    local parts = {}
-    for _, col in ipairs(st.columns) do
-      local val = data.effective_value(st, cell.row_idx, col)
-      table.insert(parts, val or "")
-    end
-    local csv = table.concat(parts, ",")
-    vim.fn.setreg("+", csv)
-    vim.notify("Yanked row as CSV", vim.log.levels.INFO)
-  end, "Yank row as CSV")
-
-  -- gY: yank table as CSV
-  kmap("grid_yank_table", function()
-    local session = ctx.session()
-    if not session then return end
-    local st = session.state
-    local r = session._render
-    if not r then return end
-    local lines_out = { table.concat(st.columns, ",") }
-    for _, row_idx in ipairs(r.ordered) do
-      local parts = {}
-      for _, col in ipairs(st.columns) do
-        local val = data.effective_value(st, row_idx, col)
-        table.insert(parts, val or "")
-      end
-      table.insert(lines_out, table.concat(parts, ","))
-    end
-    vim.fn.setreg("+", table.concat(lines_out, "\n"))
-    vim.notify("Yanked " .. #r.ordered .. " rows as CSV", vim.log.levels.INFO)
-  end, "Yank table as CSV")
-
-  -- gy: yank table as Markdown pipe table
-  kmap("grid_yank_md", function()
-    local session_gy = ctx.session()
-    if not session_gy or not session_gy._render then return end
-    local st_gy = session_gy.state
-    local r_gy = session_gy._render
-    local cols = st_gy.columns
-    local rows_data = {}
-    for _, row_idx in ipairs(r_gy.ordered) do
-      local row = {}
-      for _, col in ipairs(cols) do
-        -- effective_value returns nil for NULL; table.insert(row, nil) is a
-        -- no-op that would shift the remaining columns left, so map to "".
-        table.insert(row, data.effective_value(st_gy, row_idx, col) or "")
-      end
-      table.insert(rows_data, row)
-    end
-    vim.fn.setreg("+", table.concat(format_export(rows_data, cols, "markdown"), "\n"))
-    vim.notify("Copied as Markdown table (" .. #rows_data .. " rows)", vim.log.levels.INFO)
-  end, "Yank table as Markdown")
-
-  -- x: set cell to NULL
-  kmap("grid_null", function()
-    local session = ctx.session()
-    if not session then return end
-    if not ctx.is_editable() then
-      vim.notify("Read-only: no primary key detected", vim.log.levels.INFO)
-      return
-    end
-    local cell = M.get_cell(bufnr)
-    if not cell then
-      vim.notify("Move cursor to a data row", vim.log.levels.INFO)
-      return
-    end
-    local new_state = data.add_change(session.state, cell.row_idx, cell.col_name, nil)
-    vim.notify(cell.col_name .. " set to NULL", vim.log.levels.INFO)
-    M.apply_edit(bufnr, new_state)
-  end, "Set cell to NULL")
-
-  -- gU: multi-cursor column set (same value for all visible rows)
-  kmap("grid_column_set", function()
-    M._column_set(bufnr)
-  end, "Set column value for all visible rows")
-
-end
-
---- Visual mode: selection clamped to the data rows (gg/G/j/k) plus batch
---- edit / delete / NULL over the selected rows.
-local function setup_visual_batch_keymaps(bufnr, ctx)
-  local vmap, kvmap, get_visual_rows = ctx.vmap, ctx.kvmap, ctx.get_visual_rows
-  -- Visual gg/G/j/k: clamp selection to the data-row range so it stops at the
-  -- last data row instead of overshooting onto the separator/footer/hint line
-  -- (issue #20). Mirrors the normal-mode clamps; like them, it ignores counts.
-  local function visual_nav(target_fn)
-    return function()
-      local session = ctx.session()
-      if not session or not session._render then return end
-      local r = session._render
-      local cursor = vim.api.nvim_win_get_cursor(0)
-      local target = M._clamp_data_line(r, target_fn(r, cursor[1]))
-      vim.api.nvim_win_set_cursor(0, { target, cursor[2] })
-    end
-  end
-  vmap("gg", visual_nav(function() return 0 end), "First data row (visual)")
-  vmap("G", visual_nav(function() return math.huge end), "Last data row (visual)")
-  vmap("j", visual_nav(function(_, line) return line + 1 end), "Down, clamped to data rows")
-  vmap("k", visual_nav(function(_, line) return line - 1 end), "Up, clamped to data rows")
-
-  -- Visual e: batch edit (set all selected cells in column to same value)
-  kvmap("grid_v_edit", function()
-    local session = ctx.session()
-    if not session then return end
-    if not ctx.is_editable() then
-      vim.notify("Read-only: no primary key detected", vim.log.levels.INFO)
-      return
-    end
-    local cell = M.get_cell(bufnr)
-    if not cell then return end
-    local row_indices = get_visual_rows()
-    if not row_indices or #row_indices == 0 then return end
-    -- Exit visual mode
-    vim.api.nvim_feedkeys(vim.api.nvim_replace_termcodes("<Esc>", true, false, true), "n", false)
-    local col_name = cell.col_name
-    local enum_vals = M.enum_hint_values(session, col_name)
-    editor.open("Set " .. #row_indices .. " cells (" .. col_name .. ")", cell.value, function(new_val)
-      if new_val == nil then return end
-      local actual = editor.resolve_null(new_val)
-      local st = session.state
-      for _, ri in ipairs(row_indices) do
-        st = data.add_change(st, ri, col_name, actual)
-      end
-      M.apply_edit(bufnr, st)
-      vim.notify("Set " .. #row_indices .. " cells in " .. col_name, vim.log.levels.INFO)
-    end, { enum_values = enum_vals })
-  end, "Batch edit selected cells")
-
-  -- Visual d: toggle delete on all selected rows
-  kvmap("grid_v_delete", function()
-    local session = ctx.session()
-    if not session then return end
-    if not ctx.is_editable() then
-      vim.notify("Read-only: no primary key detected", vim.log.levels.INFO)
-      return
-    end
-    local row_indices = get_visual_rows()
-    if not row_indices or #row_indices == 0 then return end
-    vim.api.nvim_feedkeys(vim.api.nvim_replace_termcodes("<Esc>", true, false, true), "n", false)
-    local st = session.state
-    for _, ri in ipairs(row_indices) do
-      if st.inserted[ri] then
-        st = data.undo_row(st, ri)
-      else
-        st = data.toggle_delete(st, ri)
-      end
-    end
-    M.apply_edit(bufnr, st)
-  end, "Batch toggle delete")
-
-  -- Visual x: set all selected cells to NULL
-  kvmap("grid_v_null", function()
-    local session = ctx.session()
-    if not session then return end
-    if not ctx.is_editable() then
-      vim.notify("Read-only: no primary key detected", vim.log.levels.INFO)
-      return
-    end
-    local cell = M.get_cell(bufnr)
-    if not cell then return end
-    local row_indices = get_visual_rows()
-    if not row_indices or #row_indices == 0 then return end
-    vim.api.nvim_feedkeys(vim.api.nvim_replace_termcodes("<Esc>", true, false, true), "n", false)
-    local col_name = cell.col_name
-    local st = session.state
-    for _, ri in ipairs(row_indices) do
-      st = data.add_change(st, ri, col_name, nil)
-    end
-    M.apply_edit(bufnr, st)
-    vim.notify("Set " .. #row_indices .. " cells to NULL in " .. col_name, vim.log.levels.INFO)
-  end, "Batch set NULL")
-end
-
---- Inspection floats: staged SQL preview/copy, table info and properties,
---- column rename, cell explain, cell expand, row view, cell buffer, JSON
---- tree, visual stack-inspect and two-row compare.
-local function setup_inspect_keymaps(bufnr, ctx)
-  local kmap, vmap, kvmap, get_visual_rows, edit_cell, km = ctx.kmap, ctx.vmap, ctx.kvmap, ctx.get_visual_rows, ctx.edit_cell, ctx.km
-  -- gs: preview staged SQL (or pending mutation SQL) in float
-  kmap("grid_preview_sql", function()
-    local session = ctx.session()
-    if not session then return end
-
-    -- Mutation pending: show the pending SQL
-    if session.pending_mutation then
-      if session._live_sql_win and vim.api.nvim_win_is_valid(session._live_sql_win) then
-        M._close_live_sql_float(session)
-      else
-        local pm_lines = {}
-        for line in (session.pending_mutation.sql .. "\n"):gmatch("([^\n]*)\n") do
-          table.insert(pm_lines, line)
-        end
-        local grip_win = vim.api.nvim_get_current_win()
-        open_info_float(grip_win, pm_lines, { title = " " .. session.pending_mutation.type .. " SQL ", filetype = "sql" })
-      end
-      return
-    end
-
-    local st = session.state
-    if not data.has_changes(st) then
-      vim.notify("No staged changes", vim.log.levels.INFO)
-      return
-    end
-    if not st.table_name then
-      vim.notify("SQL preview requires a table name", vim.log.levels.INFO)
-      return
-    end
-    local preview = sql.preview_staged(
-      st.table_name,
-      data.get_updates(st),
-      data.get_deletes(st),
-      data.get_inserts(st)
-    )
-    local lines = {}
-    for line in (preview .. "\n"):gmatch("([^\n]*)\n") do
-      table.insert(lines, line)
-    end
-    local grip_win = vim.api.nvim_get_current_win()
-    open_info_float(grip_win, lines, { title = " Staged SQL ", filetype = "sql" })
-  end, "Preview staged SQL")
-
-  -- gc: copy staged SQL to clipboard
-  kmap("grid_copy_sql", function()
-    local session = ctx.session()
-    if not session then return end
-    local st = session.state
-    if not data.has_changes(st) then
-      vim.notify("No staged changes", vim.log.levels.INFO)
-      return
-    end
-    if not st.table_name then
-      vim.notify("SQL preview requires a table name", vim.log.levels.INFO)
-      return
-    end
-    local preview = sql.preview_staged(
-      st.table_name,
-      data.get_updates(st),
-      data.get_deletes(st),
-      data.get_inserts(st)
-    )
-    vim.fn.setreg("+", preview)
-    vim.notify("Copied SQL to clipboard", vim.log.levels.INFO)
-  end, "Copy staged SQL")
-
-  -- gi: table info float
-  kmap("grid_table_info", function()
-    local session = ctx.session()
-    if not session then return end
-    local st = session.state
-    if not st.table_name then
-      vim.notify("Table info requires a table name", vim.log.levels.INFO)
-      return
-    end
-    -- Use cached column info if available
-    if not session._column_info then
-      local info, err = db.get_column_info(st.table_name, st.url)
-      if err then
-        vim.notify("Failed to get column info: " .. err, vim.log.levels.WARN)
-        return
-      end
-      session._column_info = info
-    end
-    local info = session._column_info
-    -- Pre-compute column widths for aligned display
-    local max_name_w, max_type_w = 0, 0
-    for _, col in ipairs(info) do
-      max_name_w = math.max(max_name_w, vim.fn.strdisplaywidth(col.column_name))
-      max_type_w = math.max(max_type_w, vim.fn.strdisplaywidth(col.data_type))
-    end
-    max_name_w = math.min(max_name_w, 30)
-    max_type_w = math.min(max_type_w, 32)
-    local lines = { " " .. st.table_name, " " .. string.rep("─", 40) }
-    for _, col in ipairs(info) do
-      local name_pad = string.rep(" ", math.max(0, max_name_w - vim.fn.strdisplaywidth(col.column_name)))
-      local type_pad = string.rep(" ", math.max(0, max_type_w - vim.fn.strdisplaywidth(col.data_type)))
-      local parts = { "  " .. col.column_name .. name_pad .. "  " .. col.data_type .. type_pad }
-      if col.is_nullable == "NO" then table.insert(parts, "NOT NULL") end
-      if col.column_default ~= "" then table.insert(parts, "DEFAULT " .. col.column_default) end
-      if col.constraints ~= "" then table.insert(parts, "[" .. col.constraints .. "]") end
-      table.insert(lines, table.concat(parts, "  "))
-    end
-    local grip_win = vim.api.nvim_get_current_win()
-    open_info_float(grip_win, lines, { title = " Table Info " })
-  end, "Table info")
-
-  -- gI: full table properties float
-  kmap("grid_table_props", function()
-    local session = ctx.session()
-    if not session then return end
-    local st = session.state
-    if not st.table_name then
-      vim.notify("Table properties requires a table name", vim.log.levels.INFO)
-      return
-    end
-    local properties = require("dadbod-grip.properties")
-    local grip_win = vim.api.nvim_get_current_win()
-    properties.open(st.table_name, st.url, grip_win)
-  end, "Table properties")
-
-  -- gN: rename column under cursor
-  kmap("grid_rename_col", function()
-    local session = ctx.session()
-    if not session or not session.state.table_name then
-      vim.notify("Rename requires a table name", vim.log.levels.INFO)
-      return
-    end
-    local cell = M.get_cell(bufnr)
-    if not cell then
-      vim.notify("Move cursor to a column", vim.log.levels.INFO)
-      return
-    end
-    local ddl_mod = require("dadbod-grip.ddl")
-    ddl_mod.rename_column(session.state.table_name, cell.col_name, session.url, function()
-      if session.on_refresh then session.on_refresh(bufnr) end
-    end)
-  end, "Rename column")
-
-  -- ge: explain cell
-  kmap("grid_explain_cell", function()
-    local session = ctx.session()
-    if not session then return end
-    local st = session.state
-    local cell = M.get_cell(bufnr)
-    if not cell then
-      vim.notify("Move cursor to a data row", vim.log.levels.INFO)
-      return
-    end
-    -- Fetch column info (cached)
-    if not session._column_info and st.table_name then
-      local info, err = db.get_column_info(st.table_name, st.url)
-      if not err then session._column_info = info end
-    end
-    local col_info
-    if session._column_info then
-      for _, ci in ipairs(session._column_info) do
-        if ci.column_name == cell.col_name then col_info = ci; break end
-      end
-    end
-    local status = data.row_status(st, cell.row_idx)
-    local cell_staged = status == "modified"
-      and st.changes[cell.row_idx]
-      and st.changes[cell.row_idx][cell.col_name] ~= nil
-    local display_status
-    if status == "deleted" then
-      display_status = "deleted"
-    elseif status == "inserted" then
-      display_status = "inserted"
-    elseif cell_staged then
-      display_status = "staged"
-    else
-      display_status = "original"
-    end
-    local lines = { " " .. cell.col_name }
-    lines[#lines + 1] = " " .. string.rep("─", 30)
-    if col_info then
-      lines[#lines + 1] = "  Type: " .. col_info.data_type
-      lines[#lines + 1] = "  Nullable: " .. col_info.is_nullable
-      if col_info.column_default ~= "" then
-        lines[#lines + 1] = "  Default: " .. col_info.column_default
-      end
-      if col_info.constraints ~= "" then
-        lines[#lines + 1] = "  Constraints: " .. col_info.constraints
-      end
-    end
-    lines[#lines + 1] = "  Value: " .. (cell.value or "NULL")
-    lines[#lines + 1] = "  Status: " .. display_status
-    local grip_win = vim.api.nvim_get_current_win()
-    open_info_float(grip_win, lines, {
-      title = " Cell Info ",
-      relative = "cursor",
-      row = 1, col = 0,
-    })
-  end, "Explain cell")
-
-  -- <CR>: expand cell popup (suppressed on meta views; FK view navigates to referenced table)
-  kmap("grid_edit_enter", function()
-    local session_cr = ctx.session()
-    if not session_cr or not session_cr._render then return end
-
-    -- Meta views: no editing. FK view navigates to the referenced table; others do nothing.
-    local cv = session_cr.current_view
-    if cv and cv ~= "records" then
-      if cv == "fk" then
-        local cell = M.get_cell(bufnr)
-        if cell and cell.row_idx then
-          -- Use _meta_state: session.state was restored to records after render
-          local ms   = session_cr._meta_state
-          local row  = ms and ms.rows[cell.row_idx]
-          local cols = ms and ms.columns
-          if row and cols then
-            local dir_idx, ref_tbl_idx, col_idx
-            for i, c in ipairs(cols) do
-              if c == "direction" then dir_idx = i
-              elseif c == "ref_table" then ref_tbl_idx = i
-              elseif c == "column" then col_idx = i
-              end
-            end
-            local direction = dir_idx and (row[dir_idx] or "") or ""
-            local target
-            if direction:find("outbound", 1, true) and ref_tbl_idx then
-              target = row[ref_tbl_idx]
-            elseif direction:find("inbound", 1, true) and col_idx then
-              -- column value is "tablename.column_name"
-              target = (row[col_idx] or ""):match("^([^.]+)%.")
-            end
-            if target and target ~= "" and target ~= "(none)" then
-              require("dadbod-grip").open(target, session_cr.url)
-              return
-            end
-          end
-        end
-      end
-      -- All other meta views: Enter does nothing (no cell editing in read-only views)
-      return
-    end
-
-    local cell = M.get_cell(bufnr)
-    if cell then
-      -- Data row: edit cell (spreadsheet-style Enter to edit)
-      edit_cell()
-    else
-      -- Header/type row: detect column under cursor, show full name/type
-      local r = session_cr._render
-      local ref_bp = r.hdr_byte_positions
-      if not ref_bp then return end
-      local cols = r.visible_columns or session_cr.state.columns
-      local col_nr = vim.api.nvim_win_get_cursor(0)[2]
-      local found_col
-      for _, col in ipairs(cols) do
-        local bp = ref_bp[col]
-        if bp and col_nr >= bp.start and col_nr <= bp.finish then
-          found_col = col
-          break
-        end
-      end
-      if found_col then
-        local info = { found_col }
-        if session_cr._column_info then
-          for _, ci in ipairs(session_cr._column_info) do
-            if ci.column_name == found_col then
-              table.insert(info, "Type: " .. (ci.data_type or "unknown"))
-              if ci.is_nullable then table.insert(info, "Nullable: " .. ci.is_nullable) end
-              break
-            end
-          end
-        end
-        local grip_win = vim.api.nvim_get_current_win()
-        open_info_float(grip_win, info, {
-          title = " " .. found_col .. " ",
-          relative = "cursor",
-          row = 1, col = 0,
-        })
-      end
-    end
-  end, "Expand cell value")
-
-  -- K: row view (vertical transpose)
-  kmap("grid_row_view", function()
-    local session = ctx.session()
-    if not session then return end
-    local cell = M.get_cell(bufnr)
-    if not cell then
-      vim.notify("Move cursor to a data row", vim.log.levels.INFO)
-      return
-    end
-    local st = session.state
-    local max_name_w = 0
-    for _, col in ipairs(st.columns) do
-      max_name_w = math.max(max_name_w, vim.fn.strdisplaywidth(col))
-    end
-    local lines = {}
-    -- col_line_ranges[col] = { start=1-indexed first line, count=n lines, is_null=bool }
-    local col_line_ranges = {}
-    for _, col in ipairs(st.columns) do
-      local val = data.effective_value(st, cell.row_idx, col)
-      local pad = string.rep(" ", max_name_w - vim.fn.strdisplaywidth(col))
-      local prefix = " " .. col .. pad .. "   "
-      local start_line = #lines + 1
-      -- JSON-aware display: try to pretty-print JSON/jsonb values
-      local json_ok, json_decoded = val and pcall(vim.fn.json_decode, val) or false, nil
-      if json_ok then json_decoded = vim.fn.json_decode(val) end
-      if json_ok and type(json_decoded) == "table" then
-        local json_lines = json_to_lines(json_decoded)
-        if json_lines and #json_lines > 0 then
-          table.insert(lines, prefix .. json_lines[1])
-          for ji = 2, #json_lines do
-            table.insert(lines, string.rep(" ", #prefix) .. json_lines[ji])
-          end
-        else
-          table.insert(lines, prefix .. (val or "NULL"))
-        end
-      else
-        local display_val = val and val:gsub("\n", "↵"):gsub("\r", "") or "NULL"
-        table.insert(lines, prefix .. display_val)
-      end
-      col_line_ranges[col] = { start = start_line, count = #lines - start_line + 1, is_null = (val == nil) }
-    end
-    local grip_win = vim.api.nvim_get_current_win()
-    local _, popup_buf = open_info_float(grip_win, lines, {
-      title = " Row " .. cell.row_idx .. " ",
-    })
-    -- Shadow ]p/[p: Vim's built-in "put indented" would E21 on modifiable=false popup buffers
-    vim.keymap.set("n", "]p", "<Nop>", { buffer = popup_buf, silent = true })
-    vim.keymap.set("n", "[p", "<Nop>", { buffer = popup_buf, silent = true })
-    -- gK inside the row view: JSON tree drilldown for the column under cursor
-    local jt_key = km.get("grid_json_tree")
-    if jt_key then
-      vim.keymap.set("n", jt_key, function()
-        local cur = vim.api.nvim_win_get_cursor(0)[1]
-        for _, col in ipairs(st.columns) do
-          local range = col_line_ranges[col]
-          if range and cur >= range.start and cur < range.start + range.count then
-            require("dadbod-grip.json_tree").open(
-              data.effective_value(st, cell.row_idx, col),
-              { title = " " .. col .. " (JSON) ", origin_win = grip_win })
-            return
-          end
-        end
-      end, { buffer = popup_buf, silent = true, nowait = true })
-    end
-    -- Apply highlights using actual line ranges (not column index, which breaks on multi-line JSON)
-    local status = data.row_status(st, cell.row_idx)
-    local row_ns = vim.api.nvim_create_namespace("grip_row_view")
-    local function hl_range(range, hl_group)
-      for li = range.start, range.start + range.count - 1 do
-        vim.api.nvim_buf_set_extmark(popup_buf, row_ns, li - 1, 0, {
-          end_col = #lines[li],
-          hl_group = hl_group,
-        })
-      end
-    end
-    if status == "modified" and st.changes[cell.row_idx] then
-      for _, col in ipairs(st.columns) do
-        local range = col_line_ranges[col]
-        if range then
-          if range.is_null and st.changes[cell.row_idx][col] ~= nil then
-            -- Modified + null → orange (matches mutation preview staging color)
-            hl_range(range, "GripNullStaged")
-          elseif range.is_null then
-            hl_range(range, "GripNull")
-          elseif st.changes[cell.row_idx][col] ~= nil then
-            hl_range(range, "GripModified")
-          end
-        end
-      end
-    elseif status == "inserted" then
-      for li = 1, #lines do
-        vim.api.nvim_buf_set_extmark(popup_buf, row_ns, li - 1, 0, {
-          end_col = #lines[li],
-          hl_group = "GripInserted",
-        })
-      end
-    elseif status == "deleted" then
-      for li = 1, #lines do
-        vim.api.nvim_buf_set_extmark(popup_buf, row_ns, li - 1, 0, {
-          end_col = #lines[li],
-          hl_group = "GripDeleted",
-        })
-      end
-    else
-      -- Normal row: highlight NULL values
-      for _, col in ipairs(st.columns) do
-        local range = col_line_ranges[col]
-        if range and range.is_null then
-          hl_range(range, "GripNull")
-        end
-      end
-    end
-  end, "Row view")
-
-  -- gB: open cell value in a full split buffer (large JSON, long text)
-  kmap("grid_cell_buffer", function()
-    require("dadbod-grip.cell_buffer").open(bufnr)
-  end, "Open cell in buffer")
-
-  -- gK: JSON tree drilldown for the cell under cursor
-  kmap("grid_json_tree", function()
-    local cell = M.get_cell(bufnr)
-    if not cell then
-      vim.notify("Move cursor to a data row", vim.log.levels.INFO)
-      return
-    end
-    require("dadbod-grip.json_tree").open(cell.value, {
-      title      = " " .. cell.col_name .. " (JSON) ",
-      origin_win = vim.api.nvim_get_current_win(),
-    })
-  end, "JSON tree drilldown")
-
-  -- K (visual): stack-inspect multiple selected rows in one float.
-  -- Each row becomes a labeled block separated by a blank line.
-  -- 1 row selected → same float as normal K. N rows → stacked blocks.
-  vmap("K", function()
-    local row_indices = get_visual_rows()
-    if not row_indices or #row_indices == 0 then return end
-    local session = ctx.session()
-    if not session then return end
-    local st = session.state
-
-    local max_name_w = 0
-    for _, col in ipairs(st.columns) do
-      max_name_w = math.max(max_name_w, vim.fn.strdisplaywidth(col))
-    end
-
-    local lines = {}
-    for bi, row_idx in ipairs(row_indices) do
-      -- Header: PK=value pairs, or fallback row number
-      local pk_vals = {}
-      for _, pk in ipairs(st.pks or {}) do
-        local v = data.effective_value(st, row_idx, pk)
-        table.insert(pk_vals, pk .. "=" .. (v or "NULL"))
-      end
-      local header = #pk_vals > 0 and table.concat(pk_vals, " ") or ("Row " .. bi)
-      table.insert(lines, " \226\148\128\226\148\128 " .. header .. " " .. string.rep("\226\148\128", 28))
-
-      for _, col in ipairs(st.columns) do
-        local val = data.effective_value(st, row_idx, col)
-        local pad = string.rep(" ", max_name_w - vim.fn.strdisplaywidth(col))
-        local prefix = " " .. col .. pad .. "   "
-        local json_ok, json_decoded = val and pcall(vim.fn.json_decode, val) or false, nil
-        if json_ok then json_decoded = vim.fn.json_decode(val) end
-        if json_ok and type(json_decoded) == "table" then
-          local jlines = json_to_lines(json_decoded)
-          if jlines and #jlines > 0 then
-            table.insert(lines, prefix .. jlines[1])
-            for ji = 2, #jlines do
-              table.insert(lines, string.rep(" ", #prefix) .. jlines[ji])
-            end
-          else
-            table.insert(lines, prefix .. (val or "NULL"))
-          end
-        else
-          local display_val = val and val:gsub("\n", "↵"):gsub("\r", "") or "NULL"
-          table.insert(lines, prefix .. display_val)
-        end
-      end
-
-      if bi < #row_indices then
-        table.insert(lines, "")
-      end
-    end
-
-    local title = #row_indices == 1
-      and " Row " .. row_indices[1] .. " "
-      or  " " .. #row_indices .. " rows "
-    local grip_win = vim.api.nvim_get_current_win()
-    local _, popup_buf = open_info_float(grip_win, lines, { title = title })
-    vim.keymap.set("n", "]p", "<Nop>", { buffer = popup_buf, silent = true })
-    vim.keymap.set("n", "[p", "<Nop>", { buffer = popup_buf, silent = true })
-  end, "Stack-inspect selected rows")
-
-  -- gd (visual): diff exactly 2 selected rows, highlighting what differs.
-  -- Pure display: no DB interaction. Differing cells get GripModified / DiffAdd.
-  kvmap("grid_v_compare", function()
-    local rows = get_visual_rows()
-    if not rows or #rows ~= 2 then
-      vim.notify("Select exactly 2 rows to compare", vim.log.levels.INFO)
-      return
-    end
-    local session = ctx.session()
-    if not session then return end
-    local st = session.state
-
-    local max_name_w = 0
-    for _, col in ipairs(st.columns) do
-      max_name_w = math.max(max_name_w, vim.fn.strdisplaywidth(col))
-    end
-
-    local lines  = {}
-    local marks  = {}  -- { line_idx (0-based), hl_group }
-    for _, col in ipairs(st.columns) do
-      local a = data.effective_value(st, rows[1], col) or "NULL"
-      local b = data.effective_value(st, rows[2], col) or "NULL"
-      local pad = string.rep(" ", max_name_w - vim.fn.strdisplaywidth(col))
-      local line_a = " " .. col .. pad .. "  A  " .. a:gsub("\n", "↵")
-      local line_b = " " .. col .. pad .. "  B  " .. b:gsub("\n", "↵")
-      if a ~= b then
-        table.insert(marks, { idx = #lines,     hl = "GripModified" })
-        table.insert(marks, { idx = #lines + 1, hl = "DiffAdd" })
-      end
-      table.insert(lines, line_a)
-      table.insert(lines, line_b)
-      table.insert(lines, "")
-    end
-    -- Remove trailing blank line
-    if lines[#lines] == "" then table.remove(lines) end
-
-    local title = " Row " .. rows[1] .. " vs Row " .. rows[2] .. " "
-    local grip_win = vim.api.nvim_get_current_win()
-    local _, popup_buf = open_info_float(grip_win, lines, { title = title })
-    local ns = vim.api.nvim_create_namespace("grip_row_compare")
-    for _, m in ipairs(marks) do
-      vim.api.nvim_buf_set_extmark(popup_buf, ns, m.idx, 0, {
-        end_col = 0, end_row = m.idx + 1,
-        hl_group = m.hl, hl_eol = true,
-      })
-    end
-    vim.keymap.set("n", "]p", "<Nop>", { buffer = popup_buf, silent = true })
-    vim.keymap.set("n", "[p", "<Nop>", { buffer = popup_buf, silent = true })
-  end, "Compare two selected rows")
-end
-
---- Cursor movement and column layout: column/row navigation, first/last
---- column, hide/restore columns, width cycling, visibility picker.
-local function setup_nav_keymaps(bufnr, ctx)
-  local kmap = ctx.kmap
-  -- Shared helper: navigate to column by visible index offset.
-  -- Works on data rows, header row, and type annotation row.
-  local function nav_col(offset, use_finish)
-    local session_n = ctx.session()
-    if not session_n or not session_n._render then
-      vim.notify("nav_col: no session or render", vim.log.levels.WARN)
-      return
-    end
-    local r = session_n._render
-    local cols = r.visible_columns or session_n.state.columns
-    if #cols == 0 then
-      vim.notify("nav_col: 0 columns", vim.log.levels.WARN)
-      return
-    end
-    local cursor = vim.api.nvim_win_get_cursor(0)
-
-    -- Use current row's byte positions (handles per-row multibyte differences like ·NULL·).
-    -- Header/type/separator rows and lines past the last row fall back.
-    local ref_bp = resolve_row_bp(r, cursor[1], true)
-    if not ref_bp then
-      vim.notify("nav_col: no byte positions", vim.log.levels.WARN)
-      return
-    end
-
-    -- Determine current column index from cursor byte offset
-    local col_nr = cursor[2]
-    local current_idx = 1
-    for i, col in ipairs(cols) do
-      local bp = ref_bp[col]
-      if bp and col_nr >= bp.start and col_nr <= bp.finish then
-        current_idx = i
-        break
-      elseif bp and col_nr < bp.start then
-        current_idx = math.max(1, i - 1)
-        break
-      end
-      current_idx = i  -- past all known columns, snap to last
-    end
-
-    local target_idx
-    if offset > 0 then
-      target_idx = (current_idx % #cols) + 1
-    else
-      target_idx = current_idx == 1 and #cols or current_idx - 1
-    end
-    local target_col = cols[target_idx]
-    local bp = ref_bp[target_col]
-    if not bp then return end
-    local win = vim.api.nvim_get_current_win()
-    local target_byte = bp.start
-    if use_finish then
-      -- Land on the start of the cell's last character, not bp.finish, which can
-      -- point mid-glyph (… / ·NULL·) where the cursor cannot rest.
-      local line = vim.api.nvim_buf_get_lines(bufnr, cursor[1] - 1, cursor[1], false)[1] or ""
-      target_byte = M._cell_end_byte(line, bp.finish)
-    end
-    vim.api.nvim_win_set_cursor(win, { cursor[1], target_byte })
-    -- Reveal the target column's right edge when it runs off-screen (wide columns).
-    reveal_col_edge(win, bufnr, cursor[1], bp)
-  end
-
-  -- Tab: next column
-  kmap("grid_col_tab", function() nav_col(1) end, "Next column")
-  -- S-Tab: previous column
-  kmap("grid_col_tab_back", function() nav_col(-1) end, "Previous column")
-  -- w: next column (alias for Tab)
-  kmap("grid_col_next", function() nav_col(1) end, "Next column")
-  -- b: previous column (alias for S-Tab)
-  kmap("grid_col_prev", function() nav_col(-1) end, "Previous column")
-
-  -- gg: first data row, same column
-  kmap("grid_row_first", function()
-    local session = ctx.session()
-    if not session or not session._render then return end
-    local r = session._render
-    local ds = r.data_start or 4
-    local cursor = vim.api.nvim_win_get_cursor(0)
-    vim.api.nvim_win_set_cursor(0, { ds, cursor[2] })
-  end, "First data row")
-
-  -- G: last data row, same column
-  kmap("grid_row_last", function()
-    local session = ctx.session()
-    if not session or not session._render then return end
-    local r = session._render
-    local ds = r.data_start or 4
-    local last = ds + #r.ordered - 1
-    local cursor = vim.api.nvim_win_get_cursor(0)
-    vim.api.nvim_win_set_cursor(0, { last, cursor[2] })
-  end, "Last data row")
-
-  -- k: move up, skipping separator and type rows (jump from first data row → header/type row)
-  kmap("grid_row_up", function()
-    local session = ctx.session()
-    if not session or not session._render then
-      vim.api.nvim_feedkeys("k", "n", false)
-      return
-    end
-    local r = session._render
-    local ds = r.data_start or 4
-    local cursor = vim.api.nvim_win_get_cursor(0)
-    local line = cursor[1]
-    local col_nr = cursor[2]
-    local vis_cols = r.visible_columns or (session.state and session.state.columns) or {}
-    -- From first data row or separator: jump to nearest useful row above (type row or header)
-    if line == ds or line == ds - 1 then
-      local target_line = ds - 2  -- line 3 (type row) when ds=5, line 2 (header) when ds=4
-      -- Use type_row_byte_positions when landing on type row, else hdr_byte_positions
-      local ref_bp = (ds == 5 and r.type_row_byte_positions) or r.hdr_byte_positions
-      -- Snap to same column using first data row byte positions
-      local data_bp = r.byte_positions and r.byte_positions[1]
-      local snap = data_bp and M._snap_col(vis_cols, data_bp, col_nr)
-      local col_name = snap and snap.col_name
-      local target_col = (col_name and ref_bp and ref_bp[col_name] and ref_bp[col_name].start) or col_nr
-      vim.api.nvim_win_set_cursor(0, { target_line, target_col })
-    else
-      vim.api.nvim_feedkeys("k", "n", false)
-    end
-  end, "Move up (skip separator)")
-
-  -- j: move down, skipping separator and type rows (jump from header/type row → first data row)
-  kmap("grid_row_down", function()
-    local session = ctx.session()
-    if not session or not session._render then
-      vim.api.nvim_feedkeys("j", "n", false)
-      return
-    end
-    local r = session._render
-    local ds = r.data_start or 4
-    local cursor = vim.api.nvim_win_get_cursor(0)
-    local line = cursor[1]
-    local col_nr = cursor[2]
-    local vis_cols = r.visible_columns or (session.state and session.state.columns) or {}
-    -- From header/type/separator zone: jump directly to first data row
-    if line >= 2 and line < ds then
-      if not r.ordered or #r.ordered == 0 then
-        vim.api.nvim_feedkeys("j", "n", false)
-        return
-      end
-      -- Identify column from source row byte positions
-      local ref_bp
-      if line == 2 then
-        ref_bp = r.hdr_byte_positions
-      elseif r.type_row_byte_positions and line == ds - 2 then
-        ref_bp = r.type_row_byte_positions
-      else
-        ref_bp = r.hdr_byte_positions
-      end
-      local snap = ref_bp and M._snap_col(vis_cols, ref_bp, col_nr)
-      local col_name = snap and snap.col_name
-      -- Land on same column in first data row
-      local data_bp = r.byte_positions and r.byte_positions[1]
-      local target_col = (col_name and data_bp and data_bp[col_name] and data_bp[col_name].start) or col_nr
-      vim.api.nvim_win_set_cursor(0, { ds, target_col })
-    else
-      vim.api.nvim_feedkeys("j", "n", false)
-    end
-  end, "Move down (skip separator)")
-
-  -- ^: first column of current row (works on header, type, and data rows)
-  kmap("grid_col_first2", function()
-    local session = ctx.session()
-    if not session or not session._render then return end
-    local r = session._render
-    local cols = r.visible_columns or session.state.columns
-    local cursor = vim.api.nvim_win_get_cursor(0)
-    local bp_row = resolve_row_bp(r, cursor[1])
-    if not bp_row then return end
-    local bp = bp_row[cols[1]]
-    if not bp then return end
-    vim.api.nvim_win_set_cursor(0, { cursor[1], bp.start })
-  end, "First column")
-
-  -- 0: first column (same as ^)
-  kmap("grid_col_first", function()
-    local session = ctx.session()
-    if not session or not session._render then return end
-    local r = session._render
-    local cols = r.visible_columns or session.state.columns
-    local cursor = vim.api.nvim_win_get_cursor(0)
-    local bp_row = resolve_row_bp(r, cursor[1])
-    if not bp_row then return end
-    local bp = bp_row[cols[1]]
-    if not bp then return end
-    vim.api.nvim_win_set_cursor(0, { cursor[1], bp.start })
-  end, "First column")
-
-  -- -: hide column under cursor
-  kmap("grid_hide_col", function()
-    local session = ctx.session()
-    if not session then return end
-    local cell = M.get_cell(bufnr)
-    if not cell then
-      vim.notify("Move cursor to a column to hide", vim.log.levels.INFO)
-      return
-    end
-    if not session.hidden_columns then session.hidden_columns = {} end
-    -- Count visible columns
-    local visible = 0
-    for _, col in ipairs(session.state.columns) do
-      if not session.hidden_columns[col] then visible = visible + 1 end
-    end
-    if visible <= 1 then
-      vim.notify("Cannot hide last visible column", vim.log.levels.INFO)
-      return
-    end
-    session.hidden_columns[cell.col_name] = true
-    vim.notify("Hidden: " .. cell.col_name, vim.log.levels.INFO)
-    M.render(bufnr, session.state)
-  end, "Hide column under cursor")
-
-  -- g-: restore all hidden columns
-  kmap("grid_restore_cols", function()
-    local session = ctx.session()
-    if not session then return end
-    if not session.hidden_columns or not next(session.hidden_columns) then
-      vim.notify("No hidden columns", vim.log.levels.INFO)
-      return
-    end
-    local count = 0
-    for _ in pairs(session.hidden_columns) do count = count + 1 end
-    session.hidden_columns = {}
-    vim.notify("Restored " .. count .. " hidden column(s)", vim.log.levels.INFO)
-    M.render(bufnr, session.state)
-  end, "Restore all hidden columns")
-
-  -- =: minimize column to name width, or reset to default if already minimized
-  kmap("grid_col_width", function()
-    local session = ctx.session()
-    if not session then return end
-    -- Resolve column via data row or row-type-aware byte positions (same logic as nav_col)
-    local cell = M.get_cell(bufnr)
-    local col
-    if cell then
-      col = cell.col_name
-    else
-      local r = session._render
-      if r then
-        local cursor = vim.api.nvim_win_get_cursor(0)
-        local col_nr = cursor[2]
-        local vis_cols = r.visible_columns or (session.state and session.state.columns) or {}
-        local ref_bp = resolve_row_bp(r, cursor[1], true)
-        if ref_bp then
-          local snap = M._snap_col(vis_cols, ref_bp, col_nr)
-          if snap then col = snap.col_name end
-        end
-      end
-    end
-    if not col then
-      vim.notify("Move cursor to a column to resize", vim.log.levels.INFO)
-      return
-    end
-    if not session.col_width_overrides then session.col_width_overrides = {} end
-    if not session._col_cycle        then session._col_cycle = {} end
-
-    -- Fixed indicator budget: 3 display chars covers " ▲1"/`` ▼2" (stacked sort).
-    -- Using a fixed constant means adding/removing stacked sorts after pressing `=`
-    -- never clips the header: the budget doesn't shrink when you press `=` on
-    -- a column with only a single-arrow sort (▲ = ind_w 2) and then stack more.
-    local ind_w = 3
-
-    local cycle = session._col_cycle[col]
-    if cycle == "compact" then
-      -- compact → expanded: scan ALL rows, no MAX_COL_WIDTH cap
-      local st = session.state
-      local full_w = vim.fn.strdisplaywidth(col) + ind_w
-      if st then
-        local ordered = data.get_ordered_rows(st)
-        for _, row_idx in ipairs(ordered) do
-          local v = data.effective_value(st, row_idx, col)
-          local display = (v == nil or v == "") and NULL_DISPLAY or tostring(v)
-          full_w = math.max(full_w, vim.fn.strdisplaywidth(display))
-        end
-      end
-      session.col_width_overrides[col] = full_w
-      session._col_cycle[col] = "expanded"
-      vim.notify("Column expanded: " .. col, vim.log.levels.INFO)
-    elseif cycle == "expanded" then
-      -- expanded → auto: remove override
-      session.col_width_overrides[col] = nil
-      session._col_cycle[col] = nil
-      vim.notify("Column reset: " .. col, vim.log.levels.INFO)
-    else
-      -- auto → compact: collapse to name width only; indicator will clip, that's intentional
-      local compact_w = vim.fn.strdisplaywidth(col)
-      session.col_width_overrides[col] = compact_w
-      session._col_cycle[col] = "compact"
-      vim.notify("Column compacted: " .. col, vim.log.levels.INFO)
-    end
-    M.render(bufnr, session.state)
-  end, "Compact/expand/reset column width (cycles)")
-
-  -- gH: multi-select column visibility picker
-  kmap("grid_col_vis", function()
-    local session = ctx.session()
-    if not session then return end
-    if not session.hidden_columns then session.hidden_columns = {} end
-
-    local cols = session.state.columns
-    -- Pending is a shadow copy; only written to session on <CR>
-    local pending = {}
-    for k, v in pairs(session.hidden_columns) do pending[k] = v end
-
-    local max_w = 6  -- "[✓] " prefix = 4, plus min width
-    for _, col in ipairs(cols) do
-      if #col + 6 > max_w then max_w = #col + 6 end
-    end
-    local width  = math.min(max_w + 4, vim.o.columns - 4)
-    local height = math.min(#cols + 2, vim.o.lines - 6)
-
-    local pbuf = vim.api.nvim_create_buf(false, true)
-    vim.api.nvim_set_option_value("bufhidden", "wipe", { buf = pbuf })
-
-    local function render_lines()
-      local lines = {}
-      for _, col in ipairs(cols) do
-        local vis = not pending[col]
-        lines[#lines + 1] = (vis and "  [✓] " or "  [ ] ") .. col
-      end
-      vim.api.nvim_set_option_value("modifiable", true, { buf = pbuf })
-      vim.api.nvim_buf_set_lines(pbuf, 0, -1, false, lines)
-      vim.api.nvim_set_option_value("modifiable", false, { buf = pbuf })
-    end
-
-    render_lines()
-
-    local pwin = vim.api.nvim_open_win(pbuf, true, {
-      relative   = "editor",
-      row        = math.floor((vim.o.lines - height) / 2),
-      col        = math.floor((vim.o.columns - width) / 2),
-      width      = width,
-      height     = height,
-      style      = "minimal",
-      border     = ui.border(),
-      title      = " Columns  <Space> toggle  <CR> apply  q cancel ",
-      title_pos  = "center",
-      zindex     = 55,
-    })
-    vim.api.nvim_set_option_value("cursorline", true, { win = pwin })
-    vim.api.nvim_win_set_cursor(pwin, { 1, 0 })
-
-    local function close_picker()
-      if vim.api.nvim_win_is_valid(pwin) then vim.api.nvim_win_close(pwin, true) end
-    end
-
-    local function apply()
-      -- Guard: must keep at least one visible column
-      local visible_count = 0
-      for _, col in ipairs(cols) do
-        if not pending[col] then visible_count = visible_count + 1 end
-      end
-      if visible_count == 0 then
-        vim.notify("Cannot hide all columns", vim.log.levels.INFO)
-        return
-      end
-      session.hidden_columns = pending
-      close_picker()
-      M.render(bufnr, session.state)
-    end
-
-    local bopts = { buffer = pbuf, nowait = true }
-    vim.keymap.set("n", "<Space>", function()
-      local lnum = vim.api.nvim_win_get_cursor(pwin)[1]
-      local col = cols[lnum]
-      if not col then return end
-      if pending[col] then
-        pending[col] = nil
-      else
-        pending[col] = true
-      end
-      render_lines()
-      vim.api.nvim_win_set_cursor(pwin, { lnum, 0 })
-    end, bopts)
-    vim.keymap.set("n", "<CR>", apply, bopts)
-    vim.keymap.set("n", "q",    close_picker, bopts)
-    vim.keymap.set("n", "<Esc>", close_picker, bopts)
-
-    vim.api.nvim_create_autocmd("WinLeave", {
-      group  = _ag,
-      buffer = pbuf, once = true,
-      callback = function() close_picker() end,
-    })
-  end, "Toggle column visibility (multi-select)")
-
-  -- $: last column of current row (works on header, type, and data rows)
-  kmap("grid_col_last", function()
-    local session = ctx.session()
-    if not session or not session._render then return end
-    local r = session._render
-    local cols = r.visible_columns or session.state.columns
-    local cursor = vim.api.nvim_win_get_cursor(0)
-    local bp_row = resolve_row_bp(r, cursor[1])
-    if not bp_row then return end
-    local bp = bp_row[cols[#cols]]
-    if not bp then return end
-    local win = vim.api.nvim_get_current_win()
-    vim.api.nvim_win_set_cursor(win, { cursor[1], bp.start })
-    reveal_col_edge(win, bufnr, cursor[1], bp)
-  end, "Last column")
-
-  -- e: vim-like end-of-cell. First moves to end of current cell; when already there, advances.
-  kmap("grid_col_end", function()
-    local session_e = ctx.session()
-    if not session_e or not session_e._render then return end
-    local r = session_e._render
-    local cols = r.visible_columns or session_e.state.columns
-    if #cols == 0 then return end
-    local cursor = vim.api.nvim_win_get_cursor(0)
-    local ref_bp = resolve_row_bp(r, cursor[1], true)
-    if not ref_bp then return end
-    local col_nr = cursor[2]
-    local current_idx = 1
-    for i, col in ipairs(cols) do
-      local bp = ref_bp[col]
-      if bp and col_nr >= bp.start and col_nr <= bp.finish then
-        current_idx = i
-        break
-      elseif bp and col_nr < bp.start then
-        current_idx = math.max(1, i - 1)
-        break
-      end
-      current_idx = i
-    end
-    local current_col = cols[current_idx]
-    local bp = ref_bp[current_col]
-    if not bp then nav_col(1, true); return end
-    -- Use the start byte of the cell's last character: bp.finish can point
-    -- mid-glyph (… / ·NULL·) where the normal-mode cursor can never rest, which
-    -- would wedge `e` in place. When already at that end, advance to next column.
-    local line = vim.api.nvim_buf_get_lines(bufnr, cursor[1] - 1, cursor[1], false)[1] or ""
-    local end_col = M._cell_end_byte(line, bp.finish)
-    if cursor[2] < end_col then
-      vim.api.nvim_win_set_cursor(0, { cursor[1], end_col })
-    else
-      nav_col(1, true)
-    end
-  end, "End of current cell (then next)")
-end
-
---- Saved queries (gq), modified-row jumps ({ }), paste (p/P), visual yank,
---- live SQL float (gl) and the column type row (T).
-local function setup_misc_keymaps(bufnr, ctx)
-  local kmap, kvmap, get_visual_rows = ctx.kmap, ctx.kvmap, ctx.get_visual_rows
-  -- gq: load saved query (open query pad + picker)
-  kmap("load_saved", function()
-    local session_q = ctx.session()
-    local s_url = session_q and session_q.url
-    local query_pad = require("dadbod-grip.query_pad")
-    local saved = require("dadbod-grip.saved")
-    query_pad.open(s_url, {})
-    vim.schedule(function()
-      saved.pick(function(sql_content)
-        -- Find the query pad buffer and load SQL into it
-        for _, buf in ipairs(vim.api.nvim_list_bufs()) do
-          if vim.api.nvim_buf_get_name(buf):match("grip://query") then
-            vim.api.nvim_buf_set_lines(buf, 0, -1, false, vim.split(sql_content, "\n"))
-            vim.bo[buf].modified = false
-            break
-          end
-        end
-      end)
-    end)
-  end, "Load saved query")
-
-  -- {: previous modified/staged row
-  kmap("grid_prev_mod", function()
-    local session = ctx.session()
-    if not session or not session._render then return end
-    local r = session._render
-    local cursor = vim.api.nvim_win_get_cursor(0)
-    local ds = r.data_start or 4
-    local current_order = cursor[1] - ds + 1
-    local st = session.state
-    -- Scan backwards from current position, wrapping around
-    for offset = 1, #r.ordered do
-      local idx = current_order - offset
-      if idx < 1 then idx = idx + #r.ordered end
-      local row_idx = r.ordered[idx]
-      if data.row_status(st, row_idx) ~= "clean" then
-        vim.api.nvim_win_set_cursor(0, { ds + idx - 1, cursor[2] })
-        return
-      end
-    end
-    vim.notify("No modified rows", vim.log.levels.INFO)
-  end, "Previous modified row")
-
-  -- }: next modified/staged row
-  kmap("grid_next_mod", function()
-    local session = ctx.session()
-    if not session or not session._render then return end
-    local r = session._render
-    local cursor = vim.api.nvim_win_get_cursor(0)
-    local ds = r.data_start or 4
-    local current_order = cursor[1] - ds + 1
-    local st = session.state
-    -- Scan forward from current position, wrapping around
-    for offset = 1, #r.ordered do
-      local idx = ((current_order - 1 + offset) % #r.ordered) + 1
-      local row_idx = r.ordered[idx]
-      if data.row_status(st, row_idx) ~= "clean" then
-        vim.api.nvim_win_set_cursor(0, { ds + idx - 1, cursor[2] })
-        return
-      end
-    end
-    vim.notify("No modified rows", vim.log.levels.INFO)
-  end, "Next modified row")
-
-  -- p: paste clipboard value into cell
-  kmap("grid_paste", function()
-    local session = ctx.session()
-    if not session then return end
-    if not ctx.is_editable() then
-      vim.notify("Read-only: no primary key detected", vim.log.levels.INFO)
-      return
-    end
-    local cell = M.get_cell(bufnr)
-    if not cell then
-      vim.notify("Move cursor to a data row", vim.log.levels.INFO)
-      return
-    end
-    local clipboard = vim.fn.getreg("+")
-    if clipboard == "" then
-      vim.notify("Clipboard is empty", vim.log.levels.INFO)
-      return
-    end
-    -- Trim trailing newline from clipboard
-    clipboard = clipboard:gsub("\n$", "")
-    local new_state = data.add_change(session.state, cell.row_idx, cell.col_name, clipboard)
-    vim.notify(cell.col_name .. " = " .. clipboard:sub(1, 30), vim.log.levels.INFO)
-    M.apply_edit(bufnr, new_state)
-  end, "Paste into cell")
-
-  -- P: paste multi-line clipboard into consecutive rows (spread down)
-  kmap("grid_paste_rows", function()
-    local session = ctx.session()
-    if not session then return end
-    if not ctx.is_editable() then
-      vim.notify("Read-only: no primary key detected", vim.log.levels.INFO)
-      return
-    end
-    local cell = M.get_cell(bufnr)
-    if not cell then
-      vim.notify("Move cursor to a data row", vim.log.levels.INFO)
-      return
-    end
-    local clipboard = vim.fn.getreg("+")
-    if clipboard == "" then
-      vim.notify("Clipboard is empty", vim.log.levels.INFO)
-      return
-    end
-    -- Split clipboard by newlines into values
-    local values = {}
-    for line in (clipboard .. "\n"):gmatch("([^\n]*)\n") do
-      table.insert(values, line)
-    end
-    -- Trim trailing empty entry from final newline
-    if #values > 0 and values[#values] == "" then
-      table.remove(values)
-    end
-    if #values <= 1 then
-      -- Single value: just paste into one cell (same as p)
-      local val = values[1] or clipboard:gsub("\n$", "")
-      local new_state = data.add_change(session.state, cell.row_idx, cell.col_name, val)
-      M.apply_edit(bufnr, new_state)
-      vim.notify(cell.col_name .. " = " .. val:sub(1, 30), vim.log.levels.INFO)
-      return
-    end
-    -- Multiple values: spread into consecutive rows
-    local r = session._render
-    if not r then return end
-    local ds = r.data_start or 4
-    local cursor_line = vim.api.nvim_win_get_cursor(0)[1]
-    local start_order = cursor_line - ds + 1
-    if start_order < 1 then return end
-    local st = session.state
-    local pasted = 0
-    for i, val in ipairs(values) do
-      local order_idx = start_order + i - 1
-      if order_idx > #r.ordered then break end
-      local row_idx = r.ordered[order_idx]
-      st = data.add_change(st, row_idx, cell.col_name, val)
-      pasted = pasted + 1
-    end
-    M.apply_edit(bufnr, st)
-    vim.notify("Pasted " .. pasted .. " values into " .. cell.col_name, vim.log.levels.INFO)
-  end, "Paste multi-line into consecutive rows")
-
-  -- Visual y: yank selected cells in column (newline-separated)
-  kvmap("grid_v_yank", function()
-    local session = ctx.session()
-    if not session then return end
-    local cell = M.get_cell(bufnr)
-    if not cell then return end
-    local row_indices = get_visual_rows()
-    if not row_indices or #row_indices == 0 then return end
-    vim.api.nvim_feedkeys(vim.api.nvim_replace_termcodes("<Esc>", true, false, true), "n", false)
-    local values = {}
-    for _, ri in ipairs(row_indices) do
-      local val = data.effective_value(session.state, ri, cell.col_name)
-      table.insert(values, val or "")
-    end
-    local text = table.concat(values, "\n")
-    vim.fn.setreg("+", text)
-    vim.notify("Yanked " .. #values .. " cells from " .. cell.col_name, vim.log.levels.INFO)
-  end, "Yank selected cells in column")
-
-  -- gl: toggle live SQL preview float
-  kmap("grid_live_sql", function()
-    local session = ctx.session()
-    if not session then return end
-
-    -- Mutation preview mode: use gs instead
-    if session.pending_mutation then
-      vim.notify("gs: view mutation SQL  |  a: execute  |  U: cancel", vim.log.levels.INFO)
-      return
-    end
-
-    if not session.state.table_name then
-      vim.notify("Live SQL requires a table name", vim.log.levels.INFO)
-      return
-    end
-    session.live_sql = not session.live_sql
-    if session.live_sql then
-      vim.notify("Live SQL: ON", vim.log.levels.INFO)
-      M._update_live_sql_float(session)
-    else
-      vim.notify("Live SQL: OFF", vim.log.levels.INFO)
-      M._close_live_sql_float(session)
-    end
-  end, "Toggle live SQL")
-
-  -- T: toggle column types overlay
-  kmap("grid_type_row", function()
-    local session = ctx.session()
-    if not session then return end
-    if not session.state.table_name then
-      vim.notify("Column types requires a table name", vim.log.levels.INFO)
-      return
-    end
-    -- Fetch column info if not cached
-    if not session._column_info then
-      local info, err = db.get_column_info(session.state.table_name, session.state.url)
-      if err then
-        vim.notify("Failed to get column info: " .. err, vim.log.levels.WARN)
-        return
-      end
-      session._column_info = info
-    end
-    session.show_types = not session.show_types
-    vim.notify("Column types: " .. (session.show_types and "ON" or "OFF"), vim.log.levels.INFO)
-    M.render(bufnr, session.state)
-  end, "Toggle column types")
-end
-
---- Sorting, filtering, filter presets, export-to-file and pagination.
-local function setup_sort_filter_keymaps(bufnr, ctx)
-  local kmap = ctx.kmap
-  -- Helper: warn if pending changes, return true if user wants to proceed
-  local function confirm_discard_changes(action_name)
-    local session_c = ctx.session()
-    if not session_c then return true end
-    if not data.has_changes(session_c.state) then return true end
-    local staged = data.count_staged(session_c.state)
-    local choice = vim.fn.confirm(
-      string.format("%s will discard %d unapplied change(s). Continue?", action_name, staged),
-      "&Yes\n&Cancel", 2
-    )
-    return choice == 1
-  end
-
-  -- s: sort by column (replaces existing sort)
-  kmap("grid_sort", function()
-    local session_s = ctx.session()
-    if not session_s or not session_s.query_spec then return end
-    local cell = M.get_cell(bufnr)
-    if not cell then
-      vim.notify("Move cursor to a column to sort", vim.log.levels.INFO)
-      return
-    end
-    if not confirm_discard_changes("Sort") then return end
-    local new_spec = qmod.toggle_sort(session_s.query_spec, cell.col_name)
-    if session_s.on_requery then session_s.on_requery(bufnr, new_spec) end
-  end, "Sort by column")
-
-  -- S: add/toggle secondary sort (stacked)
-  kmap("grid_sort_stack", function()
-    local session_s = ctx.session()
-    if not session_s or not session_s.query_spec then return end
-    local cell = M.get_cell(bufnr)
-    if not cell then
-      vim.notify("Move cursor to a column to sort", vim.log.levels.INFO)
-      return
-    end
-    if not confirm_discard_changes("Sort") then return end
-    local new_spec = qmod.add_sort(session_s.query_spec, cell.col_name)
-    if session_s.on_requery then session_s.on_requery(bufnr, new_spec) end
-  end, "Add secondary sort")
-
-  -- f: quick filter by cell value
-  kmap("grid_filter_cell", function()
-    local session_f = ctx.session()
-    if not session_f or not session_f.query_spec then return end
-    local cell = M.get_cell(bufnr)
-    if not cell then
-      vim.notify("Move cursor to a cell to filter", vim.log.levels.INFO)
-      return
-    end
-    if not confirm_discard_changes("Filter") then return end
-    local new_spec = qmod.quick_filter(session_f.query_spec, cell.col_name, cell.value)
-    if session_f.on_requery then session_f.on_requery(bufnr, new_spec) end
-    local display = cell.value and (cell.col_name .. " = " .. tostring(cell.value):sub(1, 30)) or (cell.col_name .. " IS NULL")
-    vim.notify("Filtered: " .. display, vim.log.levels.INFO)
-  end, "Quick filter by cell value")
-
-  -- <C-f>: freeform WHERE clause filter
-  kmap("grid_filter_where", function()
-    local session_f = ctx.session()
-    if not session_f or not session_f.query_spec then return end
-    if not confirm_discard_changes("Filter") then return end
-    local input = ui.input({ prompt = "WHERE clause (e.g. status='x' AND amount>0): " })
-    if not input then return end
-    local new_spec = qmod.add_filter(session_f.query_spec, input)
-    if session_f.on_requery then session_f.on_requery(bufnr, new_spec) end
-  end, "Filter rows (WHERE clause)")
-
-  -- F: clear all filters
-  kmap("grid_filter_clear", function()
-    local session_f = ctx.session()
-    if not session_f or not session_f.query_spec then return end
-    if not qmod.has_filters(session_f.query_spec) then
-      vim.notify("No active filters", vim.log.levels.INFO)
-      return
-    end
-    if not confirm_discard_changes("Clear filters") then return end
-    local new_spec = qmod.clear_filters(session_f.query_spec)
-    if session_f.on_requery then session_f.on_requery(bufnr, new_spec) end
-    vim.notify("Filters cleared", vim.log.levels.INFO)
-  end, "Clear all filters")
-
-  -- gn: filter column IS NULL
-  kmap("grid_filter_null", function()
-    local session_n = ctx.session()
-    if not session_n or not session_n.query_spec then return end
-    local col_name
-    local cell = M.get_cell(bufnr)
-    if cell then
-      col_name = cell.col_name
-    else
-      local r = session_n._render
-      if r then
-        local col_nr = vim.api.nvim_win_get_cursor(0)[2]
-        local cols = r.visible_columns or session_n.state.columns
-        if r.hdr_byte_positions then
-          local snapped = M._snap_col(cols, r.hdr_byte_positions, col_nr)
-          if snapped then col_name = snapped.col_name end
-        end
-      end
-    end
-    if not col_name then
-      vim.notify("Move cursor to a column first", vim.log.levels.INFO)
-      return
-    end
-    local new_spec = qmod.quick_filter(session_n.query_spec, col_name, nil)
-    if session_n.on_requery then session_n.on_requery(bufnr, new_spec) end
-    vim.notify(col_name .. " IS NULL", vim.log.levels.INFO)
-  end, "Filter: column IS NULL")
-
-  -- gF: interactive filter builder (=, !=, >, <, LIKE, IN, BETWEEN, NULL, NOT NULL)
-  kmap("grid_filter_build", function()
-    local session_gF = ctx.session()
-    if not session_gF or not session_gF.query_spec then return end
-
-    -- Resolve column name from cursor (data row or header/type row)
-    local col_name
-    local cell = M.get_cell(bufnr)
-    if cell then
-      col_name = cell.col_name
-    else
-      local r = session_gF._render
-      if r then
-        local col_nr = vim.api.nvim_win_get_cursor(0)[2]
-        local cols = r.visible_columns or session_gF.state.columns
-        if r.hdr_byte_positions then
-          local snapped = M._snap_col(cols, r.hdr_byte_positions, col_nr)
-          if snapped then col_name = snapped.col_name end
-        end
-      end
-    end
-    if not col_name then
-      vim.notify("Move cursor to a column first", vim.log.levels.INFO)
-      return
-    end
-
-    if not confirm_discard_changes("Filter") then return end
-
-    local op = ui.input({
-      prompt = "Filter " .. col_name .. " [=, !=, >, <, LIKE, IN, BETWEEN, NULL, NOT NULL]: ",
-    })
-    if not op then return end
-    op = op:upper():match("^%s*(.-)%s*$")  -- trim + uppercase
-
-    local value
-    if op ~= "NULL" and op ~= "NOT NULL" then
-      local value_prompt = op == "LIKE" and "Value (wildcards auto-added, or type %custom%): "
-        or op == "IN" and "Values (comma-separated, e.g. 1,2,3 or alice,bob): "
-        or op == "BETWEEN" and "Range (low,high, e.g. 10,100 or 2024-01-01,2024-12-31): "
-        or "Value: "
-      -- Empty is a legal filter value (e.g. `= ''`), so it is not a cancel.
-      local val = ui.input({ prompt = value_prompt, allow_empty = true })
-      if not val then return end
-      value = val
-    end
-
-    local ok3, clause = pcall(qmod.build_filter_clause, col_name, op, value)
-    if not ok3 then
-      vim.notify("Invalid filter: " .. tostring(clause), vim.log.levels.WARN)
-      return
-    end
-
-    local new_spec = qmod.add_filter(session_gF.query_spec, clause)
-    if session_gF.on_requery then session_gF.on_requery(bufnr, new_spec) end
-    local display
-    if op == "NULL" or op == "NOT NULL" then
-      display = col_name .. " IS " .. (op == "NULL" and "" or "NOT ") .. "NULL"
-    elseif op == "BETWEEN" then
-      display = col_name .. " BETWEEN " .. tostring(value or "")
-    else
-      display = col_name .. " " .. op .. " " .. tostring(value or "")
-    end
-    vim.notify("Filtered: " .. display:sub(1, 60) .. "  (F to clear)", vim.log.levels.INFO)
-  end, "Filter builder (=, !=, >, <, LIKE, IN, BETWEEN, NULL, NOT NULL)")
-
-  -- gX: export result set to file (csv/json/sql)
-  kmap("grid_export_file", function() M.do_export(bufnr) end, "Export to file (csv/json/sql)")
-
-  -- gp: load a saved filter preset
-  kmap("grid_preset_load", function()
-    local session_fp = ctx.session()
-    if not session_fp or not session_fp.query_spec then return end
-    local tbl = session_fp.state.table_name
-    if not tbl then
-      vim.notify("Filter presets require a table context", vim.log.levels.INFO)
-      return
-    end
-    if not confirm_discard_changes("Load filter preset") then return end
-    local filters = require("dadbod-grip.filters")
-    filters.pick(tbl, function(preset)
-      local new_spec = qmod.set_filters(session_fp.query_spec, preset.clause)
-      if session_fp.on_requery then session_fp.on_requery(bufnr, new_spec) end
-      vim.notify("Filter: " .. preset.name, vim.log.levels.INFO)
-    end)
-  end, "Load filter preset")
-
-  -- gP: save current filter as preset
-  kmap("grid_preset_save", function()
-    local session_fp = ctx.session()
-    if not session_fp or not session_fp.query_spec then return end
-    local tbl = session_fp.state.table_name
-    if not tbl then
-      vim.notify("Filter presets require a table context", vim.log.levels.INFO)
-      return
-    end
-    if not qmod.has_filters(session_fp.query_spec) then
-      vim.notify("No active filters to save", vim.log.levels.INFO)
-      return
-    end
-    -- Combine all active filters into one clause
-    local clauses = {}
-    for _, f in ipairs(session_fp.query_spec.filters) do
-      table.insert(clauses, "(" .. f.clause .. ")")
-    end
-    local combined = table.concat(clauses, " AND ")
-    local name = ui.input({ prompt = "Save filter as: " })
-    if not name then return end
-    local filters = require("dadbod-grip.filters")
-    filters.save(tbl, name, combined)
-  end, "Save filter as preset")
-
-  -- ]p: next page
-  kmap("grid_next_page2", function()
-    local session_p = ctx.session()
-    if not session_p or not session_p.query_spec then return end
-    -- Check if we're on the last page
-    if session_p.total_rows then
-      local total_pages = math.max(1, math.ceil(session_p.total_rows / session_p.query_spec.page_size))
-      if session_p.query_spec.page >= total_pages then
-        vim.notify("Already on last page", vim.log.levels.INFO)
-        return
-      end
-    end
-    if not confirm_discard_changes("Page change") then return end
-    local new_spec = qmod.next_page(session_p.query_spec)
-    if session_p.on_requery then session_p.on_requery(bufnr, new_spec) end
-  end, "Next page")
-
-  -- [p: previous page
-  kmap("grid_prev_page2", function()
-    local session_p = ctx.session()
-    if not session_p or not session_p.query_spec then return end
-    if session_p.query_spec.page <= 1 then
-      vim.notify("Already on first page", vim.log.levels.INFO)
-      return
-    end
-    if not confirm_discard_changes("Page change") then return end
-    local new_spec = qmod.prev_page(session_p.query_spec)
-    if session_p.on_requery then session_p.on_requery(bufnr, new_spec) end
-  end, "Previous page")
-
-  -- ]P: jump to last page
-  kmap("grid_last_page", function()
-    local session_p = ctx.session()
-    if not session_p or not session_p.query_spec then return end
-    if not session_p.total_rows then
-      vim.notify("Total rows unknown", vim.log.levels.INFO)
-      return
-    end
-    local total_pages = math.max(1, math.ceil(session_p.total_rows / session_p.query_spec.page_size))
-    if session_p.query_spec.page >= total_pages then
-      vim.notify("Already on last page", vim.log.levels.INFO)
-      return
-    end
-    if not confirm_discard_changes("Page change") then return end
-    local new_spec = qmod.set_page(session_p.query_spec, total_pages)
-    if session_p.on_requery then session_p.on_requery(bufnr, new_spec) end
-  end, "Last page")
-
-  -- [P: jump to first page
-  kmap("grid_first_page", function()
-    local session_p = ctx.session()
-    if not session_p or not session_p.query_spec then return end
-    if session_p.query_spec.page <= 1 then
-      vim.notify("Already on first page", vim.log.levels.INFO)
-      return
-    end
-    if not confirm_discard_changes("Page change") then return end
-    local new_spec = qmod.set_page(session_p.query_spec, 1)
-    if session_p.on_requery then session_p.on_requery(bufnr, new_spec) end
-  end, "First page")
-
-  -- H/L: ergonomic page navigation (prev/next): single-key aliases for [p/]p
-  kmap("grid_prev_page", function()
-    local session_p = ctx.session()
-    if not session_p or not session_p.query_spec then return end
-    if session_p.query_spec.page <= 1 then
-      vim.notify("Already on first page", vim.log.levels.INFO)
-      return
-    end
-    if not confirm_discard_changes("Page change") then return end
-    local new_spec = qmod.prev_page(session_p.query_spec)
-    if session_p.on_requery then session_p.on_requery(bufnr, new_spec) end
-  end, "Previous page")
-
-  kmap("grid_next_page", function()
-    local session_p = ctx.session()
-    if not session_p or not session_p.query_spec then return end
-    if session_p.total_rows then
-      local total_pages = math.max(1, math.ceil(session_p.total_rows / session_p.query_spec.page_size))
-      if session_p.query_spec.page >= total_pages then
-        vim.notify("Already on last page", vim.log.levels.INFO)
-        return
-      end
-    end
-    if not confirm_discard_changes("Page change") then return end
-    local new_spec = qmod.next_page(session_p.query_spec)
-    if session_p.on_requery then session_p.on_requery(bufnr, new_spec) end
-  end, "Next page")
-
-  -- X: reset all query modifiers (sorts, filters, page)
-  kmap("grid_reset_view", function()
-    local session_x = ctx.session()
-    if not session_x or not session_x.query_spec then return end
-    local spec = session_x.query_spec
-    if #spec.sorts == 0 and #spec.filters == 0 and spec.page == 1 then
-      vim.notify("View already at defaults", vim.log.levels.INFO)
-      return
-    end
-    if not confirm_discard_changes("Reset view") then return end
-    local new_spec = qmod.reset(session_x.query_spec)
-    if session_x.on_requery then session_x.on_requery(bufnr, new_spec) end
-    vim.notify("View reset", vim.log.levels.INFO)
-  end, "Reset view (clear sort/filter/page)")
-end
-
---- Foreign key navigation: follow (gf), reverse lookup (gr), back (<C-o>).
-local function setup_fk_keymaps(bufnr, ctx)
-  local kmap = ctx.kmap
-  -- gf: navigate to FK referenced row
-  kmap("grid_fk_follow", function()
-    local session_fk = ctx.session()
-    if not session_fk or not session_fk.state.table_name then
-      vim.notify("FK navigation requires a table name", vim.log.levels.INFO)
-      return
-    end
-    local cell = M.get_cell(bufnr)
-    if not cell then
-      vim.notify("Move cursor to a cell", vim.log.levels.INFO)
-      return
-    end
-    if cell.value == nil then
-      vim.notify("NULL value: cannot follow FK", vim.log.levels.INFO)
-      return
-    end
-
-    -- Fetch FK metadata (cached per table)
-    if not session_fk.fk_cache then session_fk.fk_cache = {} end
-    local tbl = session_fk.state.table_name
-    if not session_fk.fk_cache[tbl] then
-      local fks, fk_err = db.get_foreign_keys(tbl, session_fk.state.url)
-      if fk_err then
-        vim.notify("FK lookup failed: " .. fk_err, vim.log.levels.WARN)
-        return
-      end
-      session_fk.fk_cache[tbl] = fks or {}
-    end
-
-    -- Find FK for this column
-    local fk_info
-    for _, fk in ipairs(session_fk.fk_cache[tbl]) do
-      if fk.column == cell.col_name then
-        fk_info = fk
-        break
-      end
-    end
-    if not fk_info then
-      vim.notify(cell.col_name .. " is not a foreign key", vim.log.levels.INFO)
-      return
-    end
-
-    -- Push current state to nav stack
-    if not session_fk.nav_stack then session_fk.nav_stack = {} end
-    table.insert(session_fk.nav_stack, {
-      query_spec = session_fk.query_spec,
-      state = session_fk.state,
-      table_name = tbl,
-      cursor_pos = vim.api.nvim_win_get_cursor(0),
-      total_rows = session_fk.total_rows,
-    })
-
-    -- Build query for referenced row
-    local ref_spec = qmod.new_table(fk_info.ref_table, session_fk.query_spec.page_size)
-    ref_spec = qmod.add_filter(ref_spec, sql.quote_ident(fk_info.ref_column) .. " = " .. sql.quote_value(cell.value))
-    local ref_sql = qmod.build_sql(ref_spec)
-
-    local result, err = db.query(ref_sql, session_fk.state.url)
-    if err then
-      table.remove(session_fk.nav_stack) -- pop on failure
-      vim.notify("FK query failed: " .. err, vim.log.levels.WARN)
-      return
-    end
-
-    -- Empty result: fetch columns from schema (same guard as do_refresh in init.lua)
-    if #result.columns == 0 then
-      local col_info = db.get_column_info(fk_info.ref_table, session_fk.state.url)
-      if col_info then
-        for _, ci in ipairs(col_info) do
-          table.insert(result.columns, ci.column_name)
-        end
-      end
-    end
-
-    -- Fetch PKs for referenced table
-    local pks = db.get_primary_keys(fk_info.ref_table, session_fk.state.url) or {}
-    result.primary_keys = pks
-    result.table_name = fk_info.ref_table
-    result.url = session_fk.state.url
-    result.sql = ref_sql
-
-    local new_state = data.new(result)
-    session_fk.query_spec = ref_spec
-    session_fk.total_rows = #result.rows
-    M.render(bufnr, new_state)
-    vim.notify(tbl .. "." .. cell.col_name .. " → " .. fk_info.ref_table, vim.log.levels.INFO)
-  end, "Follow FK to referenced row")
-
-  -- gr: reverse FK — jump to rows in other tables that reference this row
-  kmap("grid_fk_referencing", function()
-    M._fk_referencing(bufnr)
-  end, "Find referencing rows (reverse FK)")
-
-  -- <C-o>: go back in FK navigation stack
-  kmap("grid_fk_back", function()
-    local session_nav = ctx.session()
-    if not session_nav then return end
-    if not session_nav.nav_stack or #session_nav.nav_stack == 0 then
-      vim.notify("No FK navigation history", vim.log.levels.INFO)
-      return
-    end
-    local frame = table.remove(session_nav.nav_stack)
-    session_nav.query_spec = frame.query_spec
-    session_nav.total_rows = frame.total_rows
-    M.render(bufnr, frame.state)
-    -- Restore cursor
-    if frame.cursor_pos then
-      pcall(vim.api.nvim_win_set_cursor, 0, frame.cursor_pos)
-    end
-    vim.notify("Back to " .. (frame.table_name or "previous"), vim.log.levels.INFO)
-  end, "Go back (FK navigation)")
-end
-
---- Aggregates and column stats, table profile, clipboard export, EXPLAIN of
---- the current query and opening a URL from the cell under the cursor.
-local function setup_aggregate_keymaps(bufnr, ctx)
-  local kmap, kvmap, get_visual_rows = ctx.kmap, ctx.kvmap, ctx.get_visual_rows
-  -- ga: aggregate current column (normal: all rows, visual: selected rows).
-  -- row_indices = nil means "the whole column". Normal mode must NOT consult
-  -- '< / '>: those marks hold the *previous* visual selection, so ga silently
-  -- aggregated a stale range once the user had selected anything at all.
-  local function aggregate_column(row_indices)
-    local session_a = ctx.session()
-    if not session_a or not session_a._render then return end
-    local r = session_a._render
-    local st_a = session_a.state
-
-    -- Determine which column to aggregate from cursor position
-    local col_name
-    do
-      local cell = M.get_cell(bufnr)
-      if cell then
-        col_name = cell.col_name
-      else
-        -- Cursor may be on header/type row: resolve via byte position
-        local col_nr = vim.api.nvim_win_get_cursor(0)[2]
-        local cols = r.visible_columns or st_a.columns
-        if r.hdr_byte_positions then
-          local snapped = M._snap_col(cols, r.hdr_byte_positions, col_nr)
-          if snapped then col_name = snapped.col_name end
-        end
-      end
-      if not col_name then
-        vim.notify("Move cursor to a column first", vim.log.levels.INFO)
-        return
-      end
-    end
-
-    -- Collect values for the single column
-    local targets = row_indices or r.ordered
-    local values = {}
-    local numeric_values = {}
-    for _, row_idx in ipairs(targets) do
-      local val = data.effective_value(st_a, row_idx, col_name)
-      if val ~= nil then
-        table.insert(values, val)
-        local num = tonumber(val)
-        if num then table.insert(numeric_values, num) end
-      end
-    end
-
-    if #values == 0 then
-      vim.notify("ga: " .. col_name .. ": no values", vim.log.levels.INFO)
-      return
-    end
-
-    local agg_parts = { "ga: " .. col_name .. "  Count: " .. #values }
-    if #numeric_values > 0 then
-      local sum = 0
-      local min_v, max_v = numeric_values[1], numeric_values[1]
-      for _, n in ipairs(numeric_values) do
-        sum = sum + n
-        if n < min_v then min_v = n end
-        if n > max_v then max_v = n end
-      end
-      local avg = sum / #numeric_values
-      table.insert(agg_parts, string.format("Sum: %g", sum))
-      table.insert(agg_parts, string.format("Avg: %.2f", avg))
-      table.insert(agg_parts, string.format("Min: %g", min_v))
-      table.insert(agg_parts, string.format("Max: %g", max_v))
-    end
-
-    vim.notify(table.concat(agg_parts, "  │  "), vim.log.levels.INFO)
-  end
-
-  kmap("grid_aggregate", function() aggregate_column(nil) end, "Aggregate current column")
-  kvmap("grid_aggregate", function()
-    local rows_a = get_visual_rows()
-    if not rows_a or #rows_a == 0 then return end
-    aggregate_column(rows_a)
-  end, "Aggregate selected rows in column")
-
-  -- gS: column statistics
-  kmap("grid_col_stats", function()
-    local session_cs = ctx.session()
-    if not session_cs or not session_cs.state.table_name then
-      vim.notify("Column stats requires a table name", vim.log.levels.INFO)
-      return
-    end
-    local cell = M.get_cell(bufnr)
-    if not cell then
-      vim.notify("Move cursor to a column", vim.log.levels.INFO)
-      return
-    end
-    local tbl = session_cs.state.table_name
-    local col_q = sql.quote_ident(cell.col_name)
-    local stats_sql = string.format(
-      "SELECT COUNT(*) AS total, COUNT(DISTINCT %s) AS distinct_count, " ..
-      "COUNT(*) - COUNT(%s) AS null_count, MIN(%s) AS min_val, MAX(%s) AS max_val " ..
-      "FROM %s",
-      col_q, col_q, col_q, col_q, sql.quote_ident(tbl)
-    )
-    local result, err = db.query(stats_sql, session_cs.state.url)
-    if err then
-      vim.notify("Stats query failed: " .. err, vim.log.levels.WARN)
-      return
-    end
-    if not result or #result.rows == 0 then
-      vim.notify("No stats returned", vim.log.levels.INFO)
-      return
-    end
-    local row = result.rows[1]
-    local info = {
-      " " .. cell.col_name .. ": Column Statistics",
-      " " .. string.rep("─", 40),
-      "  Total:    " .. (row[1] or "?"),
-      "  Distinct: " .. (row[2] or "?"),
-      "  Nulls:    " .. (row[3] or "?"),
-      "  Min:      " .. (row[4] or "NULL"),
-      "  Max:      " .. (row[5] or "NULL"),
-    }
-
-    -- Try to get top 5 values
-    local top_sql = string.format(
-      "SELECT %s, COUNT(*) AS cnt FROM %s WHERE %s IS NOT NULL " ..
-      "GROUP BY %s ORDER BY cnt DESC LIMIT 5",
-      col_q, sql.quote_ident(tbl), col_q, col_q
-    )
-    local top_result = db.query(top_sql, session_cs.state.url)
-    if top_result and #top_result.rows > 0 then
-      table.insert(info, "")
-      table.insert(info, "  Top values:")
-      for _, r_top in ipairs(top_result.rows) do
-        local val = r_top[1] or "?"
-        local cnt = r_top[2] or "?"
-        table.insert(info, "    " .. tostring(val):sub(1, 30) .. "  (" .. cnt .. ")")
-      end
-    end
-
-    local grip_win = vim.api.nvim_get_current_win()
-    open_info_float(grip_win, info, { title = " Column Stats " })
-  end, "Column statistics")
-
-  -- gR: table profile report
-  kmap("grid_profile", function()
-    local session_pr = ctx.session()
-    if not session_pr or not session_pr.state.table_name then
-      vim.notify("Profile requires a table name", vim.log.levels.INFO)
-      return
-    end
-    local profile = require("dadbod-grip.profile")
-    profile.open(session_pr.state.table_name, session_pr.state.url)
-  end, "Table profile report")
-
-  -- gE: export in multiple formats
-  kmap("grid_export_clip", function()
-    local session_e = ctx.session()
-    if not session_e or not session_e._render then return end
-    local st_e = session_e.state
-    local r_e = session_e._render
-
-    local formats = { "CSV", "TSV", "JSON", "SQL INSERT", "Markdown", "Grip Table" }
-    vim.ui.select(formats, { prompt = "Export format:" }, function(choice)
-      if not choice then return end
-
-      local cols = st_e.columns
-      local rows_data = {}
-      for _, row_idx in ipairs(r_e.ordered) do
-        local row = {}
-        -- Indexed assignment: effective_value returns nil for NULL, and
-        -- table.insert(row, nil) is a no-op that would shift columns left.
-        -- Holes are intentional; consumers below iterate `for ci = 1, #cols`.
-        for ci, col in ipairs(cols) do
-          row[ci] = data.effective_value(st_e, row_idx, col)
-        end
-        table.insert(rows_data, row)
-      end
-
-      local FORMAT_IDS = {
-        ["CSV"] = "csv", ["TSV"] = "tsv", ["JSON"] = "json",
-        ["SQL INSERT"] = "sql", ["Markdown"] = "markdown", ["Grip Table"] = "grip",
-      }
-      -- Empty for a zero-row SQL INSERT export; still copied, as before.
-      local output = table.concat(
-        format_export(rows_data, cols, FORMAT_IDS[choice], st_e.table_name or "table_name"), "\n")
-
-      if output then
-        vim.fn.setreg("+", output)
-        vim.notify("Exported " .. #rows_data .. " rows as " .. choice .. " to clipboard", vim.log.levels.INFO)
-      end
-    end)
-  end, "Export in multiple formats")
-
-  -- gQ: explain current query (shortcut for :GripExplain)
-  kmap("grid_explain", function()
-    local session_x = ctx.session()
-    if not session_x then return end
-    local explain_sql
-    if session_x.query_spec then
-      explain_sql = qmod.build_sql(session_x.query_spec)
-    elseif session_x.query_sql then
-      explain_sql = session_x.query_sql
-    end
-    if not explain_sql or explain_sql == "" then
-      vim.notify("No query to explain", vim.log.levels.INFO)
-      return
-    end
-    -- Table arg form: keeps multi-line SQL in one piece (see switch_view).
-    vim.cmd({ cmd = "GripExplain", args = { explain_sql } })
-  end, "Explain current query")
-
-  -- gx: open URL in current cell (mirrors cell editor gx and Vim convention)
-  kmap("grid_url_open", function()
-    local cell = M.get_cell(bufnr)
-    if not cell or not cell.value then
-      vim.notify("No cell value", vim.log.levels.INFO)
-      return
-    end
-    local val = tostring(cell.value):match("^%s*(.-)%s*$")
-    if val:match("^https?://") or val:match("^ftp://") then
-      if vim.ui.open then
-        vim.ui.open(val)
-      elseif vim.fn.has("mac") == 1 then
-        vim.fn.jobstart({ "open", val }, { detach = true })
-      else
-        vim.fn.jobstart({ "xdg-open", val }, { detach = true })
-      end
-    else
-      vim.notify("Not a URL", vim.log.levels.INFO)
-    end
-  end, "Open URL in current cell")
-end
-
---- Everything that leaves the grid or changes the session: table diff, DDL,
---- schema browser, table/connection pickers, watch mode, write mode, open as
---- editable, query history, AI, pin and the result switcher.
-local function setup_tool_keymaps(bufnr, ctx)
-  local map, kmap = ctx.map, ctx.kmap
-  -- gD: diff against another table (picker with schema-overlap preview)
-  kmap("grid_diff", function()
-    local session_d = ctx.session()
-    if not session_d then return end
-    local st = session_d.state
-    if not st.table_name then
-      vim.notify("Diff requires a table name", vim.log.levels.INFO)
-      return
-    end
-    local db_mod = require("dadbod-grip.db")
-    local tables, err = db_mod.list_tables(st.url)
-    if not tables then
-      vim.notify("Grip: " .. (err or "failed to list tables"), vim.log.levels.ERROR)
-      return
-    end
-    if #tables == 0 then
-      vim.notify("Grip: no tables found", vim.log.levels.WARN)
-      return
-    end
-    -- Fetch source table columns once for preview comparison
-    local src_cols = db_mod.get_column_info(st.table_name, st.url) or {}
-    local src_set = {}
-    for _, col in ipairs(src_cols) do src_set[col.column_name] = true end
-
-    require("dadbod-grip.grip_picker").open({
-      title = "Diff " .. st.table_name .. " vs",
-      items = tables,
-      display = function(t)
-        local icon = t.type == "view" and "○" or "●"
-        return icon .. " " .. t.name
-      end,
-      preview = function(t)
-        local other_cols = db_mod.get_column_info(t.name, st.url) or {}
-        local other_set = {}
-        for _, col in ipairs(other_cols) do other_set[col.column_name] = true end
-        local lines = { st.table_name .. " ↔ " .. t.name, string.rep("─", 28), "" }
-        local shared, only_src, only_other = {}, {}, {}
-        for _, col in ipairs(src_cols) do
-          if other_set[col.column_name] then
-            table.insert(shared, "  = " .. col.column_name)
-          else
-            table.insert(only_src, "  - " .. col.column_name)
-          end
-        end
-        for _, col in ipairs(other_cols) do
-          if not src_set[col.column_name] then
-            table.insert(only_other, "  + " .. col.column_name)
-          end
-        end
-        if #shared > 0 then
-          table.insert(lines, "Shared (" .. #shared .. "):")
-          vim.list_extend(lines, shared)
-          table.insert(lines, "")
-        end
-        if #only_other > 0 then
-          table.insert(lines, "Only in " .. t.name .. ":")
-          vim.list_extend(lines, only_other)
-          table.insert(lines, "")
-        end
-        if #only_src > 0 then
-          table.insert(lines, "Only in " .. st.table_name .. ":")
-          vim.list_extend(lines, only_src)
-        end
-        return lines
-      end,
-      on_select = function(t)
-        require("dadbod-grip.diff").open(st.table_name, t.name, st.url)
-      end,
-    })
-  end, "Diff against table")
-
-  -- gV: show CREATE TABLE DDL in floating window
-  kmap("grid_show_ddl", function()
-    local session_v = ctx.session()
-    if not session_v then return end
-    local tbl = session_v.state.table_name
-    if not tbl then
-      vim.notify("DDL view requires a table name", vim.log.levels.INFO)
-      return
-    end
-    local url = session_v.state.url
-    local grip_win = vim.api.nvim_get_current_win()
-
-    local cols = db.get_column_info(tbl, url) or {}
-    local pks  = db.get_primary_keys(tbl, url) or {}
-    local fks  = (db.get_foreign_keys(tbl, url)) or {}
-    local idxs = (db.get_indexes(tbl, url)) or {}
-
-    local pk_set, fk_map = {}, {}
-    for _, pk in ipairs(pks) do pk_set[pk] = true end
-    for _, fk in ipairs(fks) do fk_map[fk.column] = fk end
-
-    local lines = { "CREATE TABLE " .. sql.quote_ident(tbl) .. " (" }
-    local col_lines = {}
-    for _, col in ipairs(cols) do
-      local chunk = "  " .. sql.quote_ident(col.column_name) .. " " .. (col.data_type or "TEXT")
-      if col.column_default and col.column_default ~= "" then
-        chunk = chunk .. " DEFAULT " .. col.column_default
-      end
-      if col.is_nullable == "NO" or col.is_nullable == false then
-        chunk = chunk .. " NOT NULL"
-      end
-      local remarks = {}
-      if pk_set[col.column_name] then table.insert(remarks, "PK") end
-      if fk_map[col.column_name] then
-        local fk = fk_map[col.column_name]
-        table.insert(remarks, "FK -> " .. fk.ref_table .. "." .. fk.ref_column)
-      end
-      if #remarks > 0 then chunk = chunk .. "  -- " .. table.concat(remarks, ", ") end
-      table.insert(col_lines, chunk)
-    end
-    if #pks > 0 then
-      local pk_cols = {}
-      for _, pk in ipairs(pks) do table.insert(pk_cols, sql.quote_ident(pk)) end
-      table.insert(col_lines, "  PRIMARY KEY (" .. table.concat(pk_cols, ", ") .. ")")
-    end
-    for _, fk in ipairs(fks) do
-      table.insert(col_lines, string.format(
-        "  FOREIGN KEY (%s) REFERENCES %s(%s)",
-        sql.quote_ident(fk.column), sql.quote_ident(fk.ref_table), sql.quote_ident(fk.ref_column)
-      ))
-    end
-    for i, line in ipairs(col_lines) do
-      table.insert(lines, i < #col_lines and (line .. ",") or line)
-    end
-    table.insert(lines, ");")
-
-    local non_pk_idxs = {}
-    for _, idx in ipairs(idxs) do
-      if idx.type ~= "PRIMARY" then table.insert(non_pk_idxs, idx) end
-    end
-    if #non_pk_idxs > 0 then
-      table.insert(lines, "")
-      for _, idx in ipairs(non_pk_idxs) do
-        local unique = idx.type == "UNIQUE" and "UNIQUE " or ""
-        local idx_cols = type(idx.columns) == "table"
-          and table.concat(idx.columns, ", ") or tostring(idx.columns or "")
-        table.insert(lines, string.format(
-          "CREATE %sINDEX %s ON %s (%s);",
-          unique, sql.quote_ident(idx.name or "idx"), sql.quote_ident(tbl), idx_cols
-        ))
-      end
-    end
-
-    open_info_float(grip_win, lines, { title = " DDL: " .. tbl .. " ", filetype = "sql" })
-  end, "Show CREATE TABLE DDL")
-
-  -- gb: schema browser sidebar (toggle/focus)
-  kmap("schema_browser", function()
-    local schema = require("dadbod-grip.schema")
-    local s = ctx.session()
-    -- For file-as-table sessions, pass the file path so sidebar shows column schema
-    local s_url = s and (s.file_path or s.url)
-    schema.toggle(s_url)
-  end, "Schema browser")
-
-  -- go / gT / gt: table picker
-  local function _pick_table()
-    local picker = require("dadbod-grip.picker")
-    local session = ctx.session()
-    local s_url = session and session.url
-    picker.pick_table(s_url, function(name)
-      local grip = require("dadbod-grip")
-      grip.open(name, s_url)
-    end)
-  end
-  map("go", _pick_table, "Pick table")
-  kmap("table_picker",     _pick_table, "Pick table")
-  kmap("table_picker_alt", _pick_table, "Pick table")
-
-  -- gC / <C-g>: switch database connection
-  local function _pick_connection()
-    require("dadbod-grip.connections").pick()
-  end
-  kmap("connections",     _pick_connection, "Switch connection")
-  kmap("connections_alt", _pick_connection, "Switch connection")
-
-  -- gW: toggle watch mode (auto-refresh on timer)
-  kmap("grid_watch", function()
-    local session = ctx.session()
-    if not session then return end
-    if session.watch_ms then
-      _stop_watch(bufnr)
-      vim.notify("Watch mode off", vim.log.levels.INFO)
-    else
-      local ms = (session.opts and session.opts.watch_ms) or 5000
-      _start_watch(bufnr, ms)
-      local secs = ms / 1000
-      local label = secs == math.floor(secs) and tostring(math.floor(secs)) .. "s" or tostring(secs) .. "s"
-      vim.notify("Watch mode on (" .. label .. ")", vim.log.levels.INFO)
-    end
-  end, "Toggle watch mode (auto-refresh)")
-
-  -- g!: toggle write mode (file write-back on apply)
-  kmap("grid_write_mode", function()
-    local session = ctx.session()
-    if not session then return end
-
-    local file_path = session.file_path
-    if not file_path then
-      vim.notify("Write mode only applies to local file connections", vim.log.levels.INFO)
-      return
-    end
-    if file_path:match("^https?://") then
-      vim.notify("Remote files are read-only", vim.log.levels.INFO)
-      return
-    end
-
-    if session.write_mode then
-      -- Turning OFF: warn if staged changes exist
-      local staged = session.state and (
-        next(session.state.changes or {}) or
-        next(session.state.deleted or {}) or
-        next(session.state.inserted or {})
-      )
-      if staged then
-        vim.notify("Staged changes exist. Apply (a) or undo (u) before disabling write mode.", vim.log.levels.WARN)
-        return
-      end
-      session.write_mode = false
-      _update_badge(bufnr)
-      vim.notify("Write mode off", vim.log.levels.INFO)
-    else
-      -- Turning ON: destructive-action confirm
-      local short = vim.fn.fnamemodify(file_path, ":t")
-      if not ui.confirm("Enable write mode for " .. short
-        .. "? Applying edits will overwrite the file. (y/N): ") then
-        return
-      end
-      session.write_mode = true
-      _update_badge(bufnr)
-      vim.notify("Write mode on: edits will overwrite " .. short, vim.log.levels.INFO)
-    end
-  end, "Toggle write mode (overwrite file on apply)")
-
-  -- gO: swap read-only query result to editable table
-  kmap("grid_open_edit", function()
-    local session = ctx.session()
-    if not session then return end
-    if not session.state.readonly then
-      vim.notify("Already editable: i=edit  o=insert  d=delete", vim.log.levels.INFO)
-      return
-    end
-    local grip = require("dadbod-grip")
-    local s_url = session.url
-    local current_win = vim.api.nvim_get_current_win()
-
-    -- Try to auto-detect table name (check all sources)
-    local detected = session.state.table_name
-      or (session.query_spec and session.query_spec.table_name)
-    local ambiguous = false
-
-    -- Helper: extract table name from SQL (handles quoted and unquoted identifiers)
-    local function extract_table_from_sql(sql_text)
-      local flat = sql_text:gsub("\n", " ")
-      -- Try quoted: FROM "table" or FROM `table`
-      local quoted = flat:match('[Ff][Rr][Oo][Mm]%s+"([^"]+)"')
-        or flat:match("[Ff][Rr][Oo][Mm]%s+`([^`]+)`")
-      if quoted then return quoted end
-      -- Unquoted: FROM table_name
-      return flat:match("[Ff][Rr][Oo][Mm]%s+([%w_%.]+)")
-    end
-
-    local function has_joins(sql_text)
-      return sql_text:upper():match("JOIN%s") ~= nil
-    end
-
-    -- Fallback: parse base_sql from query spec (original unwrapped SQL)
-    if not detected and session.query_spec and session.query_spec.base_sql then
-      detected = extract_table_from_sql(session.query_spec.base_sql)
-      ambiguous = has_joins(session.query_spec.base_sql)
-    end
-
-    -- Last resort: parse the wrapped query_sql (extract inner from _grip wrapper)
-    if not detected then
-      local sql_str = (session.query_sql or ""):gsub("\n", " ")
-      local inner_sql = sql_str:match("%(%s*(.-)%s*%)%s+AS%s+_grip")
-      local parse_target = inner_sql or sql_str
-      detected = extract_table_from_sql(parse_target)
-      ambiguous = ambiguous or has_joins(parse_target)
-    end
-
-    if detected and not ambiguous then
-      vim.notify("Opening " .. detected .. " as editable table", vim.log.levels.INFO)
-      grip.open(detected, s_url, { reuse_win = current_win })
-    else
-      -- Detection failed: show diagnostics so we can fix the root cause
-      vim.notify(string.format(
-        "gO: could not detect table\n table_name=%s | has_spec=%s | spec.table=%s\n spec.base_sql=%s\n query_sql=%s",
-        tostring(session.state.table_name),
-        tostring(session.query_spec ~= nil),
-        tostring(session.query_spec and session.query_spec.table_name),
-        tostring(session.query_spec and session.query_spec.base_sql and session.query_spec.base_sql:sub(1, 60)),
-        tostring((session.query_sql or ""):sub(1, 60))
-      ), vim.log.levels.WARN)
-      -- Still offer picker as fallback
-      local picker = require("dadbod-grip.picker")
-      picker.pick_table(s_url, function(name)
-        grip.open(name, s_url, { reuse_win = current_win })
-      end)
-    end
-  end, "Open as editable table")
-
-  -- gh: query history browser
-  kmap("query_history", function()
-    local hist = require("dadbod-grip.history")
-    local session_h = ctx.session()
-    local s_url = session_h and session_h.url
-    hist.pick(function(sql_content)
-      local query_pad = require("dadbod-grip.query_pad")
-      query_pad.open(s_url, { initial_sql = sql_content })
-    end)
-  end, "Query history")
-
-  -- A: AI SQL generation
-  kmap("ai", function()
-    local ai = require("dadbod-grip.ai")
-    if not ai.is_enabled() then
-      vim.notify("AI is disabled. Enable it in setup({ ai = { ... } })", vim.log.levels.INFO)
-      return
-    end
-    local session_ai = ctx.session()
-    local s_url = session_ai and session_ai.state.url
-    if not s_url then
-      s_url = require("dadbod-grip.db").get_url()
-    end
-    if not s_url then
-      vim.notify("No database connection for AI", vim.log.levels.WARN)
-      return
-    end
-    ai.ask(s_url)
-  end, "AI SQL generation")
-
-  kmap("grid_fill", function()
-    local input = ui.input({ prompt = "Rows to generate: ", default = "1" })
-    if not input then return end
-    local n = tonumber(input)
-    if not n or n < 1 then
-      vim.notify("Enter a number >= 1", vim.log.levels.INFO)
-      return
-    end
-    require("dadbod-grip").do_fill_rows(math.min(50, n))
-  end, "AI-generated staged rows (GripFill)")
-
-  -- gL: pin / unpin this result (exclude from auto-reuse by query pad)
-  kmap("grid_pin", function()
-    local session = ctx.session()
-    if not session then return end
-    if not session.pinned then
-      -- Check pinned_max cap before pinning
-      local limit = require("dadbod-grip").get_opts().pinned_max
-      if limit then
-        local count = 0
-        for _, s in pairs(M._sessions) do
-          if s.pinned then count = count + 1 end
-        end
-        if count >= limit then
-          vim.notify(string.format("Pin limit reached (%d). Unpin a result first (gL).", limit), vim.log.levels.WARN)
-          return
-        end
-      end
-      session.pinned = true
-      -- Append [pinned] to buffer name
-      local cur_name = vim.api.nvim_buf_get_name(bufnr)
-      if not cur_name:match("%[pinned%]") then
-        pcall(vim.api.nvim_buf_set_name, bufnr, cur_name .. " [pinned]")
-      end
-      vim.notify("Result pinned (gL to unpin, gJ to switch)", vim.log.levels.INFO)
-    else
-      session.pinned = false
-      -- Remove [pinned] suffix from buffer name
-      local cur_name = vim.api.nvim_buf_get_name(bufnr)
-      pcall(vim.api.nvim_buf_set_name, bufnr, cur_name:gsub("%s*%[pinned%]", ""))
-      vim.notify("Result unpinned", vim.log.levels.INFO)
-    end
-    _update_badge(bufnr)
-  end, "Pin/unpin result (exclude from auto-reuse)")
-
-  -- gJ: result switcher: pick from all live grip grid sessions
-  kmap("grid_results", function()
-    local items = {}
-    for bnum, s in pairs(M._sessions) do
-      local win_id = nil
-      for _, wid in ipairs(vim.api.nvim_tabpage_list_wins(0)) do
-        if vim.api.nvim_win_get_buf(wid) == bnum then
-          win_id = wid
-          break
-        end
-      end
-      if not win_id then goto continue end  -- bufhidden=wipe: no window means already wiped
-      local name = vim.api.nvim_buf_get_name(bnum)
-      local short = name:match("^grip://(.+)$") or name
-      local rows = s.total_rows
-      local row_label = rows and (" [" .. rows .. " rows]") or ""
-      local pin_label = s.pinned and " [pinned]" or ""
-      local elapsed = s.elapsed_ms and (" " .. s.elapsed_ms .. "ms") or ""
-      table.insert(items, {
-        bufnr   = bnum,
-        win_id  = win_id,
-        label   = short:gsub("%s*%[pinned%]", "") .. row_label .. elapsed .. pin_label,
-        pinned  = s.pinned,
-      })
-      ::continue::
-    end
-    if #items == 0 then
-      vim.notify("No open results", vim.log.levels.INFO)
-      return
-    end
-    -- Sort: pinned first, then by bufnr descending (most recent)
-    table.sort(items, function(a, b)
-      if a.pinned ~= b.pinned then return a.pinned end
-      return a.bufnr > b.bufnr
-    end)
-    require("dadbod-grip.grip_picker").open({
-      title = "Results",
-      items = items,
-      display = function(item) return (item.pinned and "" or " ") .. " " .. item.label end,
-      on_select = function(item)
-        vim.api.nvim_set_current_win(item.win_id)
-      end,
-    })
-  end, "Result switcher (all open results)")
-end
-
---- Tab view keys 1-9 plus ER diagram, help and the command palette.
-local function setup_tab_view_keymaps(bufnr, ctx)
-  local map, kmap, km = ctx.map, ctx.kmap, ctx.km
-  -- 1: schema sidebar (already in grid = always primary: open/focus sidebar)
-  kmap("tab_1", function()
-    local _s = ctx.session()
-    M.close_all_floats(_s)
-    local s_url = _s and _s.url
-    require("dadbod-grip.schema").toggle(s_url)
-  end, "Schema sidebar (key 1)")
-
-  -- 2: open query pad (pre-filled with current query)
-  kmap("tab_2", function()
-    local session_2 = ctx.session()
-    M.close_all_floats(session_2)
-    local s_url = session_2 and session_2.url
-    local initial_sql
-    if session_2 and session_2.query_spec then
-      initial_sql = qmod.clean_sql(session_2.query_spec)
-    elseif session_2 and session_2.query_sql then
-      initial_sql = session_2.query_sql
-    end
-    require("dadbod-grip.query_pad").open(s_url, initial_sql and { initial_sql = initial_sql } or nil)
-  end, "Query pad (key 2)")
-
-  -- 3: grid/records (already in non-records view = return to records; already on records = table picker)
-  kmap("tab_3", function()
-    local session_3 = ctx.session()
-    M.close_all_floats(session_3)
-    local cv = session_3 and session_3.current_view
-    if cv and cv ~= "records" then
-      M.switch_view(bufnr, "records")
-    else
-      local s_url = session_3 and session_3.url
-      local picker = require("dadbod-grip.picker")
-      picker.pick_table(s_url, function(name)
-        require("dadbod-grip").open(name, s_url)
-      end)
-    end
-  end, "Grid/records (key 3)")
-
-  -- 4-9: view tabs (4=ER diagram float, 5-9=inline grid views)
-  for n = 4, 9 do
-    local view_name = VIEW_KEYS[n]
-    local tab_key = km.get("tab_" .. n)
-    if view_name and tab_key then
-      if view_name == "er_diagram" then
-        map(tab_key, function()
-          local session = ctx.session()
-          M.close_all_floats(session)
-          local s_url = session and session.url
-          if not s_url then s_url = require("dadbod-grip.db").get_url() end
-          if not s_url then vim.notify("ER Diagram: no database connection", vim.log.levels.WARN); return end
-          require("dadbod-grip.er_diagram").toggle(s_url)
-        end, "ER diagram (key 4)")
-      else
-        map(tab_key, function()
-          M.close_all_floats(ctx.session())
-          M.switch_view(bufnr, view_name)
-        end, "View: " .. (VIEW_LABELS[view_name] or view_name))
-      end
-    end
-  end
-
-  -- gG: ER diagram float
-  kmap("er_diagram", function()
-    local session = ctx.session()
-    local s_url = session and session.url
-    if not s_url then s_url = require("dadbod-grip.db").get_url() end
-    if not s_url then
-      vim.notify("ER Diagram: no database connection", vim.log.levels.WARN)
-      return
-    end
-    local focus_table = session and session.state and session.state.table_name
-    require("dadbod-grip.er_diagram").toggle(s_url, focus_table, { focus = focus_table ~= nil })
-  end, "ER diagram (FK relationships)")
-
-  -- ?: help popup
-  kmap("help", function()
-    M.show_help({ readonly = not ctx.is_editable() })
-  end, "Show help")
-
-  -- <C-p>: command palette (discover all actions for this surface)
-  kmap("palette", function()
-    require("dadbod-grip.palette").open("grid")
-  end, "Command palette")
-end
-
---- Highlight the column under the cursor, header row included. Keyed on the
---- _render table so a redraw invalidates the cache by itself.
-local function setup_column_highlight(bufnr, ctx)
-  vim.api.nvim_create_autocmd("CursorMoved", {
-    group  = _ag,
-    buffer = bufnr,
-    callback = function()
-      local session = ctx.session()
-      if not session or not session._render then return end
-      local r = session._render
-      local vis_cols = r.visible_columns or (session.state and session.state.columns) or {}
-      if #vis_cols == 0 then return end
-
-      local cursor = vim.api.nvim_win_get_cursor(0)
-      local ref_bp = resolve_row_bp(r, cursor[1])
-      local snap = ref_bp and M._snap_col(vis_cols, ref_bp, cursor[2]) or nil
-      local col_name = snap and snap.col_name or nil
-
-      -- Nothing to redo while the cursor stays in the same column: j/k on a
-      -- 1000-row page used to clear the namespace and re-set 1001 extmarks on
-      -- every keypress. Keyed on the _render table, which render() replaces
-      -- wholesale, so a redraw invalidates this cache by itself.
-      local cached = session._col_hl
-      if cached and cached.render == r and cached.col == col_name then return end
-      session._col_hl = { render = r, col = col_name }
-
-      vim.api.nvim_buf_clear_namespace(bufnr, _col_hl_ns, 0, -1)
-      if not col_name then return end
-
-      -- Highlight header row (line 2 = index 1)
-      local hdr_bp = r.hdr_byte_positions and r.hdr_byte_positions[col_name]
-      if hdr_bp then
-        vim.api.nvim_buf_set_extmark(bufnr, _col_hl_ns, 1, hdr_bp.start, {
-          end_col = hdr_bp.finish + 1,
-          hl_group = "GripColHighlight",
-          priority = 50,
-        })
-      end
-
-      -- Highlight all data rows
-      local ds = r.data_start or 4
-      local ordered = r.ordered or {}
-      for i = 1, #ordered do
-        local bp_row = r.byte_positions and r.byte_positions[i]
-        if bp_row then
-          local bp = bp_row[col_name]
-          if bp then
-            vim.api.nvim_buf_set_extmark(bufnr, _col_hl_ns, ds + i - 2, bp.start, {
-              end_col = bp.finish + 1,
-              hl_group = "GripColHighlight",
-              priority = 50,
-            })
-          end
-        end
-      end
-    end,
-  })
-end
-
 --- Wire all buffer-local grid keymaps and the column-highlight autocmd.
 --- Section order is load bearing (see the note above make_keymap_ctx).
 function M._setup_keymaps(bufnr)
   local ctx = make_keymap_ctx(bufnr)
-  setup_edit_keymaps(bufnr, ctx)
-  setup_visual_batch_keymaps(bufnr, ctx)
-  setup_inspect_keymaps(bufnr, ctx)
-  setup_nav_keymaps(bufnr, ctx)
-  setup_misc_keymaps(bufnr, ctx)
-  setup_sort_filter_keymaps(bufnr, ctx)
-  setup_fk_keymaps(bufnr, ctx)
-  setup_aggregate_keymaps(bufnr, ctx)
-  setup_tool_keymaps(bufnr, ctx)
-  setup_tab_view_keymaps(bufnr, ctx)
-  setup_column_highlight(bufnr, ctx)
+  for _, section in ipairs(KEYMAP_SECTIONS) do
+    section.setup(bufnr, ctx)
+  end
 end
 
 --- Open the full help popup. Called from grid, query pad, and schema sidebar.
